% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{abbreviate_street_types}
\alias{abbreviate_street_types}
\title{Replace full street type words with abbreviations}
\usage{
abbreviate_street_types(addr)
}
\arguments{
\item{addr}{character vector of addresses}
}
\value{
character vector with abbreviated street types
}
\description{
Replace full street type words with abbreviations
}
\keyword{internal}
