% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{agg_enr_column_order}
\alias{agg_enr_column_order}
\title{Helper function to return aggregate enrollment columns in correct order}
\usage{
agg_enr_column_order(df)
}
\arguments{
\item{df}{aggregate enrollment dataframe}
}
\value{
data.frame
}
\description{
Helper function to return aggregate enrollment columns in correct order
}
