% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{agg_enr_pct_total}
\alias{agg_enr_pct_total}
\title{Helper function to support calculating pct_total on aggregate enr dataframes}
\usage{
agg_enr_pct_total(df)
}
\arguments{
\item{df}{aggregate enrollment dataframe}
}
\value{
data.frame
}
\description{
Helper function to support calculating pct_total on aggregate enr dataframes
}
