% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_pop_calcs.R
\name{agg_spec_pop_column_order}
\alias{agg_spec_pop_column_order}
\title{Helper function to return aggregate special populations columns in
correct order}
\usage{
agg_spec_pop_column_order(df)
}
\arguments{
\item{df}{aggregate special populations df}
}
\value{
data_frame
}
\description{
Helper function to return aggregate special populations columns in
correct order
}
