% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_calcs.R
\name{agg_sped_column_order}
\alias{agg_sped_column_order}
\title{Helper function to return aggregate sped columns in correct order}
\usage{
agg_sped_column_order(df)
}
\arguments{
\item{df}{aggregate sped dataframe}
}
\value{
data.frame
}
\description{
Helper function to return aggregate sped columns in correct order
}
