% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_enr_aggs}
\alias{allpublic_enr_aggs}
\title{Calculate All Public Enrollment Aggregates}
\usage{
allpublic_enr_aggs(df)
}
\arguments{
\item{df}{a tidied enrollment dataframe, eg output of `fetch_enr`}
}
\value{
dataframe with all public school option aggregates for any charter host city
}
\description{
Calculate All Public Enrollment Aggregates
}
