% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_gcount_aggs}
\alias{allpublic_gcount_aggs}
\title{Calculate Charter Sector Grad Count aggregates}
\usage{
allpublic_gcount_aggs(df)

allpublic_gcount_aggs(df)
}
\arguments{
\item{df}{tidied postsecondary matriculation dataframe, eg output 
of enrich_matric_counts()}
}
\value{
df containing charter sector aggregate grad count

df containing all-public sector postsecondary matriculation rates
}
\description{
Calculate Charter Sector Grad Count aggregates

Calculate All Public matriculation aggregates
}
