% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_grate_aggs}
\alias{allpublic_grate_aggs}
\title{Calculate All Public Grad Rate aggregates}
\usage{
allpublic_grate_aggs(df)
}
\arguments{
\item{df}{tidied grate dataframe, eg output of fetch_grate}
}
\value{
df containing allpublic aggregate grad rate
}
\description{
Calculate All Public Grad Rate aggregates
}
