% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_matric_aggs}
\alias{allpublic_matric_aggs}
\title{Calculate All Public matriculation aggregates}
\usage{
allpublic_matric_aggs(df)
}
\arguments{
\item{df}{tidied postsecondary matriculation dataframe, eg output 
of enrich_matric_counts()}
}
\value{
df containing all-public sector postsecondary matriculation rates
}
\description{
Calculate All Public matriculation aggregates
}
