% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_parcc_aggs}
\alias{allpublic_parcc_aggs}
\title{Calculate All Public Options PARCC Aggregates}
\usage{
allpublic_parcc_aggs(df)
}
\arguments{
\item{df}{tidied PARCC dataframe, eg output of fetch_parcc}
}
\value{
df containing all public option aggregates by district
}
\description{
Calculate All Public Options PARCC Aggregates
}
