% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_spec_pop_aggs}
\alias{allpublic_spec_pop_aggs}
\title{Calculate All City Special Populations aggregates}
\usage{
allpublic_spec_pop_aggs(df)
}
\arguments{
\item{df}{tidied grad count dataframe, eg output of 
`fetch_reportcard_special_pop``}
}
\value{
df with all city aggregates per host city
}
\description{
Calculate All City Special Populations aggregates
}
