% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_sped_aggs}
\alias{allpublic_sped_aggs}
\title{Calculate Charter Sector SPED aggregates}
\usage{
allpublic_sped_aggs(df)
}
\arguments{
\item{df}{tidied grad count dataframe, eg output of fetch_sped}
}
\value{
df containing charter sector aggregate sped
}
\description{
Calculate Charter Sector SPED aggregates
}
