% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{ALLPUBLIC_SUFFIX}
\alias{ALLPUBLIC_SUFFIX}
\title{Suffix for all public aggregations}
\format{
An object of class \code{character} of length 1.
}
\usage{
ALLPUBLIC_SUFFIX
}
\description{
Suffix for all public aggregations
}
\keyword{internal}
