% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{arrange_enr}
\alias{arrange_enr}
\title{Arrange enrollment file columns}
\usage{
arrange_enr(df)
}
\arguments{
\item{df}{Cleaned enrollment file}
}
\value{
Data frame with columns in standard order
}
\description{
Put an enrollment file in the correct column order.
}
\keyword{internal}
