% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped.R
\name{build_sped_url}
\alias{build_sped_url}
\title{Build SPED data URL}
\usage{
build_sped_url(end_year)
}
\arguments{
\item{end_year}{ending school year}
}
\value{
URL string for the SPED data file
}
\description{
Build SPED data URL
}
\keyword{internal}
