% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if a key exists in cache}
\usage{
cache_exists(key)
}
\arguments{
\item{key}{Cache key}
}
\value{
Logical
}
\description{
Check if a key exists in cache
}
\keyword{internal}
