% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get}
\alias{cache_get}
\title{Get data from cache}
\usage{
cache_get(key)
}
\arguments{
\item{key}{Cache key}
}
\value{
Cached data or NULL if not found
}
\description{
Get data from cache
}
\keyword{internal}
