% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set}
\alias{cache_set}
\title{Store data in cache}
\usage{
cache_set(key, value)
}
\arguments{
\item{key}{Cache key}

\item{value}{Data to cache}
}
\value{
The value (invisibly)
}
\description{
Store data in cache
}
\keyword{internal}
