% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_percentiles.R
\name{calc_nj_percentiles}
\alias{calc_nj_percentiles}
\title{Calculate NJ Percentiles}
\usage{
calc_nj_percentiles(tidy_df)
}
\arguments{
\item{tidy_df}{a tidy data frame, eg output of `tidy_nj_assess()`.}
}
\value{
the tidy data frame with two new columns - proficiency_percentile
and scale_score_percentile - describing the school/district's position
in the distribution of NJ schools/districts for that:
assessment program (eg NJASK), test name (eg ELA, Math), testing year,
grade, and testing subgroup
}
\description{
Calculate NJ Percentiles
}

