% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{charter_city}
\alias{charter_city}
\title{Charter School Host District Mapping}
\format{
A data frame with 137 rows and 6 columns:
\describe{
  \item{district_id}{Charter school district identifier}
  \item{district_name}{Charter school district name}
  \item{host_county_id}{County code of the host district}
  \item{host_county_name}{County name of the host district}
  \item{host_district_id}{Host district identifier}
  \item{host_district_name}{Host district name}
}
}
\source{
NJ Department of Education
}
\usage{
charter_city
}
\description{
A dataset containing charter schools and their host districts/cities in New Jersey.
Charter schools are public schools that operate independently but are accountable to
their authorizing districts (host districts).
}
\keyword{datasets}
