% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter_sector_enr_aggs}
\alias{charter_sector_enr_aggs}
\title{Calculate Charter Sector Enrollment Aggregates}
\usage{
charter_sector_enr_aggs(df)
}
\arguments{
\item{df}{a tidied enrollment dataframe, eg output of `fetch_enr`}
}
\value{
dataframe with charter sector aggregates per host city
}
\description{
Calculate Charter Sector Enrollment Aggregates
}
