% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter_sector_grate_aggs}
\alias{charter_sector_grate_aggs}
\title{Calculate Charter Sector Grad Rate aggregates}
\usage{
charter_sector_grate_aggs(df)
}
\arguments{
\item{df}{tidied grate dataframe, eg output of fetch_grate}
}
\value{
df containing charter sector aggregate grad rate
}
\description{
Calculate Charter Sector Grad Rate aggregates
}
