% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter_sector_matric_aggs}
\alias{charter_sector_matric_aggs}
\title{Calculate Charter Sector postsecondary matriculation aggregates}
\usage{
charter_sector_matric_aggs(df)
}
\arguments{
\item{df}{tidied matriculation dataframe, eg output of 
enrich_matric_counts}
}
\value{
df containing charter sector matriculation aggregates
}
\description{
Calculate Charter Sector postsecondary matriculation aggregates
}
