% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter_sector_parcc_aggs}
\alias{charter_sector_parcc_aggs}
\title{Calculate Charter Sector PARCC aggregates}
\usage{
charter_sector_parcc_aggs(df)
}
\arguments{
\item{df}{tidied PARCC dataframe, eg output of fetch_parcc}
}
\value{
df containing charter sector aggregate PARCC performance
}
\description{
Calculate Charter Sector PARCC aggregates
}
