% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter_sector_spec_pop_aggs}
\alias{charter_sector_spec_pop_aggs}
\title{Calculate Charter Sector Special Populations Aggregates}
\usage{
charter_sector_spec_pop_aggs(df)
}
\arguments{
\item{df}{a tidied enrollment dataframe, eg output of 
`fetch_reportcard_special_pop`}
}
\value{
dataframe with charter sector aggregates per host city
}
\description{
Calculate Charter Sector Special Populations Aggregates
}
