% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter_sector_sped_aggs}
\alias{charter_sector_sped_aggs}
\title{Calculate Charter Sector SPED Aggregates}
\usage{
charter_sector_sped_aggs(df)
}
\arguments{
\item{df}{a tidied district special education dataframe, 
e.g. output of `fetch_sped`}
}
\value{
dataframe with charter sector aggregates per host city
}
\description{
Calculate Charter Sector SPED Aggregates
}
