% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{CHARTER_SECTOR_SUFFIX}
\alias{CHARTER_SECTOR_SUFFIX}
\title{Suffix for charter sector aggregations}
\format{
An object of class \code{character} of length 1.
}
\usage{
CHARTER_SECTOR_SUFFIX
}
\description{
Suffix for charter sector aggregations
}
\keyword{internal}
