% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{clean_enr_data}
\alias{clean_enr_data}
\title{Clean enrollment data types}
\usage{
clean_enr_data(df)
}
\arguments{
\item{df}{An enrollment data frame with standardized column names}
}
\value{
Data frame with correct data types
}
\description{
All columns come back as character; coerce some back to numeric.
}
\keyword{internal}
