% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{clean_enr_grade}
\alias{clean_enr_grade}
\title{Tidy up the grade level field on enrollment data}
\usage{
clean_enr_grade(df)
}
\arguments{
\item{df}{An enrollment data file. clean_enr_grade is part of a set
of chained cleaning functions that live inside process_enr.}
}
\value{
Data frame with cleaner grade_level column
}
\description{
Tidy up the grade level field on enrollment data
}
\keyword{internal}
