% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enr.R
\name{clean_enr_names}
\alias{clean_enr_names}
\title{clean enrollment names}
\usage{
clean_enr_names(df)
}
\arguments{
\item{df}{a enr data frame (eg output of \code{get_raw_enr})}
}
\description{
give consistent names to the enrollment files
}

