% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_graduation.R
\name{clean_grate_names}
\alias{clean_grate_names}
\title{Clean graduation rate names}
\usage{
clean_grate_names(name_vector)
}
\arguments{
\item{name_vector}{Vector of subgroup names}
}
\value{
Vector of cleaned subgroup names
}
\description{
Standardizes subgroup names in graduation data.
}
\keyword{internal}
