% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_id}
\alias{clean_id}
\title{clean_id}
\usage{
clean_id(x)
}
\arguments{
\item{x}{id vector}
}
\value{
cleaned id vector
}
\description{
cleans district id columns(mixed numeric/character) of leading zeros
}
