% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_names.R
\name{clean_name}
\alias{clean_name}
\title{Internal helper to map a column name}
\usage{
clean_name(df_names, clean_list)
}
\arguments{
\item{df_names}{Column name to clean}

\item{clean_list}{Named list mapping old names to new names}
}
\value{
Cleaned column name
}
\description{
Internal helper to map a column name
}
\keyword{internal}
