% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{clean_rc_enrollment}
\alias{clean_rc_enrollment}
\title{Clean Report Card Enrollment}
\usage{
clean_rc_enrollment(list_of_dfs)
}
\arguments{
\item{list_of_dfs}{output of get_one_rc_database (ie, a list  of data.frames)}
}
\value{
data frame with school enrollment
}
\description{
Clean Report Card Enrollment
}
