% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{clean_spr_subgroups}
\alias{clean_spr_subgroups}
\title{Clean SPR subgroup names}
\usage{
clean_spr_subgroups(group)
}
\arguments{
\item{group}{Vector of subgroup names}
}
\value{
Vector of cleaned subgroup names
}
\description{
Standardizes subgroup names from SPR database to match
package naming conventions.
}
\keyword{internal}
