% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_percentiles.R
\name{dfg_peer_percentile}
\alias{dfg_peer_percentile}
\title{Calculate DFG peer percentile by grade}
\usage{
dfg_peer_percentile(df)
}
\arguments{
\item{df}{}
}
\value{
data.frame with percent proficient and scale score percentile rank
}
\description{
calculates DFG percentile by grade/test
}
