% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matric_calcs.R
\name{district_matric_aggs}
\alias{district_matric_aggs}
\title{Aggregates matriculation data by district. Only school-level matriculation
data reported before 2017. This function approximates district level 
results. If schools within the district do not report for certain subgroups,
the approximation will be further off.}
\usage{
district_matric_aggs(df)
}
\arguments{
\item{df}{output of \code{enrich_matric_counts}}
}
\value{
A data frame of district aggregations
}
\description{
Aggregates matriculation data by district. Only school-level matriculation
data reported before 2017. This function approximates district level 
results. If schools within the district do not report for certain subgroups,
the approximation will be further off.
}
