% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{district_name_to_id}
\alias{district_name_to_id}
\title{Given friendly names, returns ids}
\usage{
district_name_to_id(district_names, df)
}
\arguments{
\item{district_names}{vector or list of friendly district names}

\item{df}{df used to generate the friendly names}
}
\value{
vector of district_ids
}
\description{
Given friendly names, returns ids
}
