% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_college_career.R
\docType{data}
\name{.biliteracy_summary_num_cols}
\alias{.biliteracy_summary_num_cols}
\title{Numeric columns to coerce on each redesigned biliteracy sheet}
\format{
An object of class \code{character} of length 10.
}
\usage{
.biliteracy_summary_num_cols
}
\description{
Numeric columns to coerce on each redesigned biliteracy sheet
}
\keyword{internal}
