% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_enrollment.R
\name{.parse_enr_archive}
\alias{.parse_enr_archive}
\title{Parse a validated zipped fall enrollment artifact}
\usage{
.parse_enr_archive(tname, end_year)
}
\arguments{
\item{tname}{Path to a validated NJ DOE enrollment ZIP.}

\item{end_year}{A school year. Year is the end of the academic year - eg
2006-07 school year is year '2007'.}
}
\value{
Data frame with raw enrollment data
}
\description{
Parse a validated zipped fall enrollment artifact
}
\keyword{internal}
