% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{ELEMENTARY_GRADES}
\alias{ELEMENTARY_GRADES}
\title{Elementary grades (K-5)}
\format{
An object of class \code{character} of length 6.
}
\usage{
ELEMENTARY_GRADES
}
\description{
Elementary grades (K-5)
}
\keyword{internal}
