% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ell.R
\docType{data}
\name{ELL_TIDY_COLS}
\alias{ELL_TIDY_COLS}
\title{Canonical column order for tidy EL population data}
\format{
An object of class \code{character} of length 22.
}
\usage{
ELL_TIDY_COLS
}
\description{
Canonical column order for tidy EL population data
}
\keyword{internal}
