% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{enr_aggs}
\alias{enr_aggs}
\title{Calculate enrollment aggregates}
\usage{
enr_aggs(df)
}
\arguments{
\item{df}{Cleaned enrollment dataframe, eg output of `clean_enr_data`}
}
\value{
Data frame with aggregated columns
}
\description{
Aggregates gender columns into racial subgroups (e.g., white_m + white_f = white)
}
\keyword{internal}
