% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation_aggs.R
\name{enrich_grad_count}
\alias{enrich_grad_count}
\title{Enrich report card matriculation percentages with grad counts}
\usage{
enrich_grad_count(df, end_year)
}
\arguments{
\item{df}{Data frame including subgroup percentages}

\item{end_year}{Numeric end year of grad counts to join}
}
\value{
data_frame with graduated_count and cohort_count columns added
}
\description{
Joins graduation count data to a data frame containing subgroup percentages.
}
