% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{enrich_matric_counts}
\alias{enrich_matric_counts}
\title{Enrich matriculation rates with counts from grad counts}
\usage{
enrich_matric_counts(df, type = "16 month")
}
\arguments{
\item{df}{data frame of including matriculation percentages}

\item{type}{character string specifying matriculation type. One of "16 month" or "12 month"}
}
\value{
data_frame
}
\description{
Enrich matriculation rates with counts from grad counts
}
