% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{enrich_school_city_ward}
\alias{enrich_school_city_ward}
\title{Enrich School Data with City Ward}
\usage{
enrich_school_city_ward(df)
}
\arguments{
\item{df}{any dataframe with a district_id}
}
\value{
df enriched with ward, if geographic data is 'registered' for a given district
}
\description{
Enrich School Data with City Ward
}
