% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{expand_street_types}
\alias{expand_street_types}
\title{Replace abbreviated street types with full words}
\usage{
expand_street_types(addr)
}
\arguments{
\item{addr}{character vector of addresses}
}
\value{
character vector with expanded street types
}
\description{
Replace abbreviated street types with full words
}
\keyword{internal}
