% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{extract_pr_SAT}
\alias{extract_pr_SAT}
\title{Extract Progress Report SAT School Averages}
\usage{
extract_pr_SAT(list_of_prs)
}
\arguments{
\item{list_of_prs}{output of get_rc_databases (ie, a list where each element is)
a list of data.frames}
}
\value{
data frame with all years of SAT School Averages present in the input
}
\description{
Extract Progress Report SAT School Averages
}
