% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{extract_rc_cds}
\alias{extract_rc_cds}
\title{Extract Report Card CDS}
\usage{
extract_rc_cds(list_of_prs)
}
\arguments{
\item{list_of_prs}{output of get_rc_databases (ie, a list where each element is)
a list of data.frames}
}
\value{
data frame with county, district and school ids and identifiers for all 
years present in the input
}
\description{
Extract Report Card CDS
}
