% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{extract_rc_college_matric}
\alias{extract_rc_college_matric}
\title{Extract Report Card Matriculation Rates}
\usage{
extract_rc_college_matric(list_of_prs, school_only = TRUE,
  cds_identifiers = TRUE)
}
\value{
data frame with all the years of 4 year and 2 year college matriculation data
present in in the input
}
\description{
Extract Report Card Matriculation Rates
}
