% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{extract_rc_enrollment}
\alias{extract_rc_enrollment}
\title{Extract Report Card Enrollment}
\usage{
extract_rc_enrollment(list_of_prs, cds_identifiers = TRUE)
}
\arguments{
\item{list_of_prs}{output of get_rc_databases (ie, a list where each element is
a list of data.frames)}

\item{cds_identifiers}{add the county, district and school name?  default is TRUE}
}
\value{
data frame with school enrollment
}
\description{
Extract Report Card Enrollment
}
