% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{extract_rc_SAT}
\alias{extract_rc_SAT}
\title{Extract Report Car SAT School Averages}
\usage{
extract_rc_SAT(list_of_prs, school_only = TRUE)
}
\arguments{
\item{list_of_prs}{output of get_rc_databases (ie, a list where each element is)
a list of data.frames}

\item{school_only}{some years have district average results, not just school-level. 
if school_only, return only school data.  default is TRUE}
}
\value{
data frame with all years of SAT School Averages present in the input
}
\description{
Extract Report Car SAT School Averages
}
