% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{facilities_capture}
\alias{facilities_capture}
\title{Internal: first regex capture or NA}
\usage{
facilities_capture(text, pattern, group = 1L)
}
\description{
Internal: first regex capture or NA
}
\keyword{internal}
