% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{facilities_capture_vec}
\alias{facilities_capture_vec}
\title{Internal vectorized regex capture}
\usage{
facilities_capture_vec(x, pattern, group = 1L)
}
\description{
Internal vectorized regex capture
}
\keyword{internal}
