% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_facilities.R
\name{facilities_columns}
\alias{facilities_columns}
\title{Canonical facilities long-schema columns}
\usage{
facilities_columns()
}
\value{
Character vector of canonical column names.
}
\description{
Canonical facilities long-schema columns
}
\keyword{internal}
