% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{facilities_empty_long}
\alias{facilities_empty_long}
\title{Internal empty facilities long frame before NCES attachment}
\usage{
facilities_empty_long()
}
\description{
Internal empty facilities long frame before NCES attachment
}
\keyword{internal}
