% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{facilities_html_text}
\alias{facilities_html_text}
\title{Internal: very small HTML-to-text helper}
\usage{
facilities_html_text(html)
}
\description{
Internal: very small HTML-to-text helper
}
\keyword{internal}
