% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{facilities_melt}
\alias{facilities_melt}
\title{Internal: source-aware melt helper}
\usage{
facilities_melt(
  df,
  entity_id_col,
  entity_name_col,
  entity_level,
  src,
  spec,
  county_id_col = NA_character_,
  district_id_col = NA_character_,
  school_id_col = NA_character_
)
}
\description{
Internal: source-aware melt helper
}
\keyword{internal}
