% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{facilities_million_to_usd}
\alias{facilities_million_to_usd}
\title{Internal: convert strings like 87.2 Million to whole-dollar text}
\usage{
facilities_million_to_usd(x)
}
\description{
Internal: convert strings like 87.2 Million to whole-dollar text
}
\keyword{internal}
