% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{facilities_number_string}
\alias{facilities_number_string}
\title{Internal: remove commas from a number-like string}
\usage{
facilities_number_string(x)
}
\description{
Internal: remove commas from a number-like string
}
\keyword{internal}
