% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_facility_gis.R
\name{facility_gis_layers}
\alias{facility_gis_layers}
\title{Internal: shipped New Jersey facilities GIS layers}
\usage{
facility_gis_layers()
}
\description{
Internal: shipped New Jersey facilities GIS layers
}
\keyword{internal}
