% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_course_enrollment.R
\name{fetch_arts_enrollment}
\alias{fetch_arts_enrollment}
\title{Fetch Visual and Performing Arts Enrollment Data}
\usage{
fetch_arts_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with arts course enrollment by discipline
  (Music, Visual Art, Theater, Dance, etc.)
}
\description{
Downloads visual and performing arts course participation data from SPR database.
}
\examples{
\dontrun{
arts <- fetch_arts_enrollment(2024)

# View music enrollment
arts \%>\%
  dplyr::filter(grepl("Music|Band|Choir", course_type)) \%>\%
  dplyr::select(school_name, number_of_students)
}
}
