% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_course_enrollment.R
\name{fetch_cs_enrollment}
\alias{fetch_cs_enrollment}
\title{Fetch Computer Science Enrollment Data}
\usage{
fetch_cs_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with computer science course enrollment by course type
  (AP CS A, AP CS Principles, Intro to CS, etc.)
}
\description{
Downloads computer science course participation data from SPR database.
}
\examples{
\dontrun{
cs <- fetch_cs_enrollment(2024)

# View AP Computer Science enrollment
cs \%>\%
  dplyr::filter(grepl("AP", course_type)) \%>\%
  dplyr::select(school_name, number_of_students)
}
}
