% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_graduation.R
\name{fetch_grad_count}
\alias{fetch_grad_count}
\title{Fetch Grad Counts}
\usage{
fetch_grad_count(end_year)
}
\arguments{
\item{end_year}{End of the academic year - eg 2006-07 is 2007.
Valid values are 2012-2024.}
}
\value{
dataframe with grad counts
}
\description{
Downloads and processes graduation count data.
}
\examples{
\dontrun{
# Get 2023 graduation counts
gcount_2023 <- fetch_grad_count(2023)
}
}
