% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grate.R
\name{fetch_grate}
\alias{fetch_grate}
\title{fetch grate}
\usage{
fetch_grate(end_year, tidy = TRUE)
}
\arguments{
\item{end_year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 1998-2014.}

\item{tidy}{tidy the response into a data frame with consistent headers?}
}
\value{
a
}
\description{
a consistent interface into the NJ HS graduates data.
}

