% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa.R
\name{fetch_hspa}
\alias{fetch_hspa}
\title{gets and processes a HSPA file}
\usage{
fetch_hspa(end_year)
}
\arguments{
\item{end_year}{a school end_year.  end_year is the end of the academic year - eg 2013-14
school year is end_year '2014'.  valid values are 2004-2014.}
}
\description{
\code{fetch_njask} is a wrapper around \code{get_raw_hspa} and
\code{process_nj_assess} that passes the correct file layout data to each function,
given a end_year and grade.
}
