% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_college_career.R
\name{fetch_ib_participation}
\alias{fetch_ib_participation}
\title{Fetch IB Participation Data}
\usage{
fetch_ib_participation(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024). Year is the end of the academic
year - eg 2020-21 school year is end_year '2021'.}

\item{level}{One of "school" or "district". "school" returns school-level
data, "district" returns district and state-level data.}
}
\value{
Data frame with IB participation metrics
}
\description{
Downloads and extracts International Baccalaureate participation from the
SPR database. Note: Most IB data is included in the AP/IB sheet, use
\code{\link{fetch_ap_participation}} for comprehensive data.
}
\examples{
\dontrun{
ib <- fetch_ib_participation(2024)
}
}
