% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_course_enrollment.R
\name{fetch_science_course_enrollment}
\alias{fetch_science_course_enrollment}
\title{Fetch Science Course Enrollment Data}
\usage{
fetch_science_course_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with science course enrollment by subject area (Biology,
  Chemistry, Physics, Environmental Science, etc.)
}
\description{
Downloads science course participation data from SPR database.
}
\examples{
\dontrun{
science <- fetch_science_course_enrollment(2024)

# View physics enrollment
science \%>\%
  dplyr::filter(course_type == "Physics") \%>\%
  dplyr::select(school_name, subgroup, number_of_students)
}
}
