% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{fetch_staff_demographics}
\alias{fetch_staff_demographics}
\title{Fetch Staff Demographics Data}
\usage{
fetch_staff_demographics(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with staff race/gender breakdowns
}
\description{
Downloads teacher/administrator demographic data from SPR database.
}
\examples{
\dontrun{
staff <- fetch_staff_demographics(2024)
}
}
