% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_course_enrollment.R
\name{fetch_world_language_enrollment}
\alias{fetch_world_language_enrollment}
\title{Fetch World Language Enrollment Data}
\usage{
fetch_world_language_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with world language course enrollment by language
  (Spanish, French, German, Italian, Chinese, etc.)
}
\description{
Downloads world languages course participation data from SPR database.
}
\examples{
\dontrun{
world_lang <- fetch_world_language_enrollment(2024)

# View Spanish enrollment
world_lang \%>\%
  dplyr::filter(grepl("Spanish", course_type)) \%>\%
  dplyr::select(school_name, number_of_students)
}
}
