% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{format_valid_values}
\alias{format_valid_values}
\title{Format valid values for error messages}
\usage{
format_valid_values(values, max_show = 10)
}
\arguments{
\item{values}{Vector of valid values}

\item{max_show}{Maximum number of values to show}
}
\value{
Character string with formatted values
}
\description{
Format valid values for error messages
}
\keyword{internal}
