% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{friendly_district_names}
\alias{friendly_district_names}
\title{Friendly District Names}
\usage{
friendly_district_names(df)
}
\arguments{
\item{df}{data.frame that contains district_id and district_name}
}
\value{
character vector of names, one per district id
}
\description{
Friendly District Names
}
\keyword{internal}
