% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{friendly_school_names}
\alias{friendly_school_names}
\title{Friendly School Names}
\usage{
friendly_school_names(df)
}
\arguments{
\item{df}{data.frame that contains school_id and school_name}
}
\value{
character vector of names, one per school id
}
\description{
Friendly School Names
}
\keyword{internal}
