% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate_calcs.R
\name{gcount_aggregate_calcs}
\alias{gcount_aggregate_calcs}
\title{Aggregate multiple grad count rows and produce summary statistics}
\usage{
gcount_aggregate_calcs(df)
}
\arguments{
\item{df}{grouped df of gcount data}
}
\value{
df with aggregate stats for whatever grouping was provided
}
\description{
Aggregate multiple grad count rows and produce summary statistics
}
