% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{gcount_column_order}
\alias{gcount_column_order}
\title{Grad Count column order}
\usage{
gcount_column_order(df)
}
\arguments{
\item{df}{processsed grad count df}
}
\value{
df in correct order
}
\description{
Grad Count column order
}
