% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\docType{data}
\name{GCOUNT_VALID_YEARS}
\alias{GCOUNT_VALID_YEARS}
\title{Graduation count valid years}
\format{
An object of class \code{integer} of length 27.
}
\usage{
GCOUNT_VALID_YEARS
}
\description{
Graduation count valid years
}
\keyword{internal}
