% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_access_url}
\alias{get_access_url}
\title{Get the URL for ACCESS for ELLs data file}
\usage{
get_access_url(end_year)
}
\arguments{
\item{end_year}{A school year (2022-2024)}
}
\value{
URL string
}
\description{
Builds the URL for a given year's ACCESS data file.
URL structure changed between years.
}
\keyword{internal}
