% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_registry.R
\name{get_available_ell_years}
\alias{get_available_ell_years}
\title{Get English Learner population years}
\usage{
get_available_ell_years()
}
\value{
Integer vector of school-year end years.
}
\description{
Get English Learner population years
}
