% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_registry.R
\name{get_available_years}
\alias{get_available_years}
\title{Get enrollment years available from the tidy front door}
\usage{
get_available_years()
}
\value{
Integer vector of school-year end years.
}
\description{
Get enrollment years available from the tidy front door
}
