% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_chronic_absenteeism_url}
\alias{get_chronic_absenteeism_url}
\title{Get the URL for ESSA Accountability Workbook containing chronic absenteeism data}
\usage{
get_chronic_absenteeism_url(end_year)
}
\arguments{
\item{end_year}{A school year (2017-2024, excluding 2020-2021)}
}
\value{
URL string
}
\description{
Builds the URL for a given year's accountability workbook file.
}
\keyword{internal}
