% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_names.R
\name{get_enr_column_order}
\alias{get_enr_column_order}
\title{Standard enrollment column order}
\usage{
get_enr_column_order()
}
\value{
Character vector of column names in standard order
}
\description{
Standard enrollment column order
}
\keyword{internal}
