% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_names.R
\name{get_enr_types}
\alias{get_enr_types}
\title{Enrollment column type specifications}
\usage{
get_enr_types()
}
\value{
Named list of column types
}
\description{
Enrollment column type specifications
}
\keyword{internal}
