% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R
\name{get_enr_url}
\alias{get_enr_url}
\title{Enrollment file URL configuration}
\usage{
get_enr_url(end_year)
}
\arguments{
\item{end_year}{The school year (end year)}
}
\value{
Character string URL
}
\description{
Returns the URL for enrollment data for a given year.
}
\keyword{internal}
