% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essa.R
\name{get_essa_comprehensive_file}
\alias{get_essa_comprehensive_file}
\title{Get the ESSA comprehensive file}
\usage{
get_essa_comprehensive_file(end_year)
}
\arguments{
\item{end_year}{a school year.  end_year is the end of the academic year - eg 2016-17
school year is end_year 2017.  valid values are 2017}
}
\value{
list of data frames
}
\description{
Get the ESSA comprehensive file
}
