% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essa.R
\name{get_essa_file}
\alias{get_essa_file}
\title{Get an ESSA comprehensive or targeted accountability file}
\usage{
get_essa_file(end_year, file_type = "comprehensive")
}
\arguments{
\item{end_year}{a school year.  end_year is the end of the academic year - eg 2016-17
school year is end_year 2017.  valid values are 2017}

\item{file_type}{'comprehensive' or 'targeted'?}
}
\value{
list of data frames
}
\description{
Get an ESSA comprehensive or targeted accountability file
}
