% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{get_grad_count}
\alias{get_grad_count}
\title{Get NJ graduation count data}
\usage{
get_grad_count(end_year)
}
\arguments{
\item{end_year}{end of the academic year - eg 2006-07 is 2007.
valid values are 1998-present.}
}
\value{
dataframe with the number of graduates per school and district
}
\description{
Get NJ graduation count data
}
