% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_graduation.R
\name{get_grad_rate}
\alias{get_grad_rate}
\title{Get NJ graduation rate data}
\usage{
get_grad_rate(end_year, methodology)
}
\arguments{
\item{end_year}{End of the academic year - 2011-2012 is 2012.
Valid values are 2011-2024.}

\item{methodology}{Character string specifying calculation methodology.
One of "4 year" or "5 year".}
}
\value{
dataframe with the number of graduates per school and district
}
\description{
Get NJ graduation rate data
}
\keyword{internal}
