% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R
\name{get_grad_url}
\alias{get_grad_url}
\title{Get graduation data URL}
\usage{
get_grad_url(end_year, methodology = "4 year")
}
\arguments{
\item{end_year}{Graduation cohort year}

\item{methodology}{Either "4 year" or "5 year"}
}
\value{
Character string URL
}
\description{
Get graduation data URL
}
\keyword{internal}
