% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{get_merged_rc_database}
\alias{get_merged_rc_database}
\title{Combines school and district Performance Reports for 2017-on, when two files were released.}
\usage{
get_merged_rc_database(end_year)
}
\arguments{
\item{end_year}{end of the academic year.  Valid values are 2017, 2018, 2019}
}
\value{
list of dataframes
}
\description{
Combines school and district Performance Reports for 2017-on, when two files were released.
}
