% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R
\name{get_parcc_url}
\alias{get_parcc_url}
\title{Get PARCC/NJSLA assessment URL}
\usage{
get_parcc_url(end_year, grade, subj)
}
\arguments{
\item{end_year}{Assessment year}

\item{grade}{Grade level or course code}

\item{subj}{Subject ("ela" or "math")}
}
\value{
Character string URL
}
\description{
Get PARCC/NJSLA assessment URL
}
\keyword{internal}
