% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_percentiles.R
\name{get_percentile_cols}
\alias{get_percentile_cols}
\title{Get percentile cols}
\usage{
get_percentile_cols(df)
}
\arguments{
\item{df}{data.frame, output of `assessment_peer_percentile`}
}
\value{
slim df with limited columns
}
\description{
internal/helper function to facilitate the extraction of relevant 
`assessment_peer_percentile` columns when calculating state/dfg-wide 
percentiles
}
