% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_raw_access}
\alias{get_raw_access}
\title{Reads the raw ACCESS for ELLs Excel file from the state website}
\usage{
get_raw_access(end_year, grade = "all")
}
\arguments{
\item{end_year}{A school year (2022-2024)}

\item{grade}{Grade level: "K" or 0 for Kindergarten, or 1-12 for other grades.
Use "all" to get all grades combined.}
}
\value{
ACCESS dataframe for the specified grade
}
\description{
Downloads the ACCESS file and reads a specific grade sheet.
}
\keyword{internal}
