% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gepa.R
\name{get_raw_gepa}
\alias{get_raw_gepa}
\title{read a fixed width, raw GEPA data file from the NJ state website}
\usage{
get_raw_gepa(year, layout = layout_gepa)
}
\arguments{
\item{year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 2004-2007.}

\item{layout}{what layout dataframe to use.  default is layout_gepa.}
}
\description{
\code{get_raw_gepa} builds a url and uses readr's \code{read_fwf} to get the fixed
width text file into a R data frame
}

