% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{get_raw_grate}
\alias{get_raw_grate}
\title{read a zipped excel HS graduate outcome/rate file from the NJ state website}
\usage{
get_raw_grate(end_year, calc_type = "4 year")
}
\arguments{
\item{end_year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 1998-2014.}

\item{calc_type}{c('4 year', '5 year').  5 year only available for 2012 and 2013 
as of 8/26/15}
}
\description{
\code{get_raw_grate} returns a data frame with NJ HS grad rate
}

