% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hspa.R
\name{get_raw_hspa}
\alias{get_raw_hspa}
\title{read a fixed width, raw HSPA data file from the NJ state website}
\usage{
get_raw_hspa(year, layout = layout_hspa)
}
\arguments{
\item{year}{a school year.  year is the end of the academic year - eg 2013-14
school year is year '2014'.  valid values are 2004-2014.}

\item{layout}{what layout dataframe to use.  default is layout_hspa.}
}
\description{
\code{get_raw_hspa} builds a url and uses readr's \code{read_fwf} to get the fixed
width text file into a R data frame
}

