% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njask.R
\name{get_raw_njask}
\alias{get_raw_njask}
\title{read a fixed width, raw NJASK data file from the NJ state website}
\usage{
get_raw_njask(end_year, grade, layout = layout_njask)
}
\arguments{
\item{end_year}{a school year.  end_year is the end of the academic year - eg 2013-14
school year is end_year '2014'.  valid values are 2004-2014.}

\item{grade}{a grade level.  valid values are 3,4,5,6,7,8}

\item{layout}{what layout dataframe to use.  default is layout_njask.}
}
\description{
\code{get_raw_njask} builds a url and uses readr's \code{read_fwf} to get the fixed 
width text file into a R data frame
}

