% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{get_raw_njdoe_cds}
\alias{get_raw_njdoe_cds}
\title{Internal: fetch and read the NJDOE CDS workbook}
\usage{
get_raw_njdoe_cds(src)
}
\description{
Internal: fetch and read the NJDOE CDS workbook
}
\keyword{internal}
