% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{get_raw_njdoe_lead_soa}
\alias{get_raw_njdoe_lead_soa}
\title{Internal: read NJDOE lead SOA workbook, excluding early childcare rows}
\usage{
get_raw_njdoe_lead_soa(src)
}
\description{
Internal: read NJDOE lead SOA workbook, excluding early childcare rows
}
\keyword{internal}
