% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{get_raw_njdoe_sda_allocation}
\alias{get_raw_njdoe_sda_allocation}
\title{Internal: read NJDOE SDA district allocation workbook}
\usage{
get_raw_njdoe_sda_allocation(src)
}
\description{
Internal: read NJDOE SDA district allocation workbook
}
\keyword{internal}
