% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_raw_njgpa}
\alias{get_raw_njgpa}
\title{Reads raw NJGPA data from the state website}
\usage{
get_raw_njgpa(end_year, subj)
}
\arguments{
\item{end_year}{School-year end year.}

\item{subj}{Subject.}
}
\value{
NJGPA data frame.
}
\description{
Reads raw NJGPA data from the state website
}
\keyword{internal}
