% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{get_raw_njsda_active_projects}
\alias{get_raw_njsda_active_projects}
\title{Internal: extract active NJSDA project detail pages}
\usage{
get_raw_njsda_active_projects(src)
}
\description{
Internal: extract active NJSDA project detail pages
}
\keyword{internal}
