% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_raw_parcc}
\alias{get_raw_parcc}
\title{Reads raw PARCC data from the state website}
\usage{
get_raw_parcc(end_year, grade_or_subj, subj)
}
\arguments{
\item{end_year}{School-year end year.}

\item{grade_or_subj}{Grade or course.}

\item{subj}{Subject.}
}
\value{
PARCC data frame.
}
\description{
Reads raw PARCC data from the state website
}
\keyword{internal}
