% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_raw_sla}
\alias{get_raw_sla}
\title{Reads raw NJSLA data from the state website}
\usage{
get_raw_sla(end_year, grade_or_subj, subj)
}
\arguments{
\item{end_year}{School-year end year.}

\item{grade_or_subj}{Grade or course.}

\item{subj}{Subject.}
}
\value{
NJSLA data frame.
}
\description{
Reads raw NJSLA data from the state website
}
\keyword{internal}
