% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_pop.R
\name{get_reportcard_special_pop}
\alias{get_reportcard_special_pop}
\title{Get Special Population Enrollment from Report Card Files}
\usage{
get_reportcard_special_pop(end_year)
}
\arguments{
\item{end_year}{ending academic year.  valid values are 2017, 2018, 2019}
}
\value{
data.frame with special population enrollment percentages
}
\description{
Get Special Population Enrollment from Report Card Files
}
\keyword{internal}
