% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_registry.R
\name{get_school_directory_url}
\alias{get_school_directory_url}
\title{Current Homeroom school-directory CSV endpoint}
\usage{
get_school_directory_url()
}
\value{
Character URL.
}
\description{
Current Homeroom school-directory CSV endpoint
}
\keyword{internal}
