% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_graduation.R
\name{get_spr_6yr_grad_url}
\alias{get_spr_6yr_grad_url}
\title{Get SPR database URL for 6-year graduation rates}
\usage{
get_spr_6yr_grad_url(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2021-2024). Year is the end of the academic
year - eg 2020-21 school year is end_year '2021'.}

\item{level}{One of "school" or "district". Determines which database file
to download.}
}
\value{
URL string
}
\description{
Builds the URL for the School Performance Reports database containing
6-year graduation cohort profile data.
}
\keyword{internal}
