% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{get_spr_url}
\alias{get_spr_url}
\title{Get SPR Database URL}
\usage{
get_spr_url(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024). Year is the end of the academic
year - eg 2020-21 school year is end_year '2021'.}

\item{level}{One of "school" or "district". Determines which database file
to download.}
}
\value{
URL string
}
\description{
Builds the URL for the School Performance Reports database containing
multiple sheets of school performance data.
}
\keyword{internal}
