% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped.R
\name{get_valid_sped_years}
\alias{get_valid_sped_years}
\title{Valid years for SPED data}
\usage{
get_valid_sped_years()
}
\value{
vector of valid end years for SPED data
}
\description{
Valid years for SPED data
}
\keyword{internal}
