% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\name{grad_file_group_cleanup}
\alias{grad_file_group_cleanup}
\title{Grad file group cleanup}
\usage{
grad_file_group_cleanup(group)
}
\arguments{
\item{group}{Column of group (subgroup) data from NJ grad file}
}
\value{
Column with cleaned up subgroup names
}
\description{
Standardizes subgroup names across years.
}
\keyword{internal}
