% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\docType{data}
\name{GRATE_4YR_VALID_YEARS}
\alias{GRATE_4YR_VALID_YEARS}
\title{4-year graduation rate valid years}
\format{
An object of class \code{integer} of length 14.
}
\usage{
GRATE_4YR_VALID_YEARS
}
\description{
4-year graduation rate valid years
}
\keyword{internal}
