% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\docType{data}
\name{GRATE_5YR_VALID_YEARS}
\alias{GRATE_5YR_VALID_YEARS}
\title{5-year graduation rate valid years}
\format{
An object of class \code{integer} of length 8.
}
\usage{
GRATE_5YR_VALID_YEARS
}
\description{
5-year graduation rate valid years
}
\keyword{internal}
