% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_calcs.R
\name{grate_aggregate_calcs}
\alias{grate_aggregate_calcs}
\title{Aggregate multiple grad rate rows and produce summary statistics}
\usage{
grate_aggregate_calcs(df)
}
\arguments{
\item{df}{grouped df of grate data}
}
\value{
df with aggregate stats for whatever grouping was provided
}
\description{
Aggregate multiple grad rate rows and produce summary statistics
}
