% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{grate_column_order}
\alias{grate_column_order}
\title{Grad Rate column order}
\usage{
grate_column_order(df)
}
\arguments{
\item{df}{processed grad rate df}
}
\value{
df in correct order
}
\description{
Grad Rate column order
}
