% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_rank.R
\name{grate_percentile_rank}
\alias{grate_percentile_rank}
\title{Graduation rate percentile rank}
\usage{
grate_percentile_rank(
  df,
  peer_type = "statewide",
  custom_ids = NULL,
  by_subgroup = TRUE,
  by_methodology = TRUE
)
}
\arguments{
\item{df}{Output of \code{fetch_grad_rate()} or similar graduation data}

\item{peer_type}{Character. Peer group type. See \code{define_peer_group()}.}

\item{custom_ids}{Character vector. Custom peer group district IDs.}

\item{by_subgroup}{Logical. Calculate separate percentiles by subgroup?
Default TRUE.}

\item{by_methodology}{Logical. Calculate separate percentiles by methodology
(4-year, 5-year)? Default TRUE.}
}
\value{
df with grad_rate_rank, grad_rate_n, grad_rate_percentile columns
}
\description{
Convenience wrapper for calculating percentile rank of
graduation rates within a peer group.
}
