% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{HIGH_SCHOOL_GRADES}
\alias{HIGH_SCHOOL_GRADES}
\title{High school grades (9-12)}
\format{
An object of class \code{character} of length 4.
}
\usage{
HIGH_SCHOOL_GRADES
}
\description{
High school grades (9-12)
}
\keyword{internal}
