% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{id_enr_aggs}
\alias{id_enr_aggs}
\title{Identify enrollment aggregation levels}
\usage{
id_enr_aggs(df)
}
\arguments{
\item{df}{Enrollment dataframe, output of tidy_enr}
}
\value{
data.frame with boolean aggregation flags
}
\description{
Adds boolean flags to identify state, county, district, and school level records.
}
