% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{id_grad_aggs}
\alias{id_grad_aggs}
\title{Identify enrollment aggregation levels}
\usage{
id_grad_aggs(df)
}
\arguments{
\item{df}{enrollment dataframe, output of tidy_enr}
}
\value{
data.frame with boolean aggregation flags
}
\description{
Identify enrollment aggregation levels
}
