% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{id_rc_aggs}
\alias{id_rc_aggs}
\title{Identify reportcard aggregation levels}
\usage{
id_rc_aggs(df)
}
\arguments{
\item{df}{enrollment dataframe, output of extract_rc_*}
}
\value{
data.frame with boolean aggregation flags
}
\description{
Identify reportcard aggregation levels
}
\keyword{internal}
