% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{id_selected_districtids}
\alias{id_selected_districtids}
\title{District Names to IDs}
\usage{
id_selected_districtids(district_names, lookup_df)
}
\arguments{
\item{district_names}{vector of names}

\item{lookup_df}{dataframe with district_id and district_name}
}
\value{
list of districtids matching the names
}
\description{
District Names to IDs
}
