% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\name{is_district_total}
\alias{is_district_total}
\title{Check if a school is a district total}
\usage{
is_district_total(school_id)
}
\arguments{
\item{school_id}{School ID to check}
}
\value{
Logical
}
\description{
Check if a school is a district total
}
\keyword{internal}
