% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\name{is_state_aggregate}
\alias{is_state_aggregate}
\title{Check if a record is a state aggregate}
\usage{
is_state_aggregate(county_id, district_id)
}
\arguments{
\item{county_id}{County ID}

\item{district_id}{District ID}
}
\value{
Logical
}
\description{
Check if a record is a state aggregate
}
\keyword{internal}
