% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{K12_GRADES}
\alias{K12_GRADES}
\title{K-12 grades (excluding pre-K)}
\format{
An object of class \code{character} of length 13.
}
\usage{
K12_GRADES
}
\description{
K-12 grades (excluding pre-K)
}
\keyword{internal}
