% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{KINDER_PROGRAM_CODES}
\alias{KINDER_PROGRAM_CODES}
\title{Kindergarten program codes}
\format{
An object of class \code{character} of length 4.
}
\usage{
KINDER_PROGRAM_CODES
}
\description{
Kindergarten program codes
}
\keyword{internal}
