% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/layout_data.R
\docType{data}
\name{layout_hspa}
\alias{layout_hspa}
\title{HSPA Fixed-Width File Layout}
\format{
A data frame with 559 rows and 10 columns:
\describe{
  \item{field_start_position}{Starting position of field in fixed-width file}
  \item{field_end_position}{Ending position of field in fixed-width file}
  \item{field_length}{Length of field in characters}
  \item{data_type}{Data type of field (e.g., "Text", "Numeric")}
  \item{description}{Description of field}
  \item{comments}{Additional comments about field}
  \item{valid_values}{Valid values or ranges for field}
  \item{spanner1}{First-level column grouping label}
  \item{spanner2}{Second-level column grouping label}
  \item{final_name}{Final column name to use in parsed data}
}

\describe{
\item{field_start_position}{field_start_position}
\item{field_end_position}{field_end_position}
\item{field_length}{field_length}
\item{data_type}{data_type}
\item{description}{description}
\item{comments}{comments}
\item{valid_values}{valid_values}
\item{spanner1}{spanner1}
\item{spanner2}{spanner2}
\item{final_name}{final_name}
#' }
}
\source{
NJ Department of Education HSPA file specifications

NJDOE website - eg the 'File Layout' link on
http://www.nj.gov/education/schools/achievement/14/hspa/
}
\usage{
layout_hspa

layout_hspa
}
\description{
Column layout specification for reading High School Proficiency Assessment (HSPA)
fixed-width format data files.

A processed R version of the 'File Layout' for the 2011-present HSPA
Original file is in .xls format on NJDOE website
}
\keyword{datasets}
