% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/layout_data.R
\docType{data}
\name{layout_njask06gr5}
\alias{layout_njask06gr5}
\title{NJASK 2006 Grade 5 Fixed-Width File Layout}
\format{
A data frame with column layout specifications including field positions,
  lengths, data types, descriptions, and final column names.

\describe{
\item{field_start_position}{field_start_position}
\item{field_end_position}{field_end_position}
\item{field_length}{field_length}
\item{data_type}{data_type}
\item{description}{description}
\item{comments}{comments}
\item{valid_values}{valid_values}
\item{spanner1}{spanner1}
\item{spanner2}{spanner2}
\item{final_name}{final_name}
#' }
}
\source{
NJ Department of Education NJASK 2006 Grade 5 file specifications

NJDOE website - eg the 'File Layout' link on
http://www.state.nj.us/education/schools/achievement/2007/njask57/
}
\usage{
layout_njask06gr5

layout_njask06gr5
}
\description{
Column layout specification for reading New Jersey Assessment of Skills and Knowledge
(NJASK) 2006 Grade 5 fixed-width format data files.

A processed R version of the 'File Layout' for the 2006
NJASK, middle school (gr 5, 6, 7) version.   
Original file is in .xls format on NJDOE website
}
\keyword{datasets}
