% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_data.R
\docType{data}
\name{layout_njask09}
\alias{layout_njask09}
\title{NJASK layout metadata, 2009 test, 2007-8 gr 5+ test}
\format{
\describe{
\item{field_start_position}{field_start_position}
\item{field_end_position}{field_end_position}
\item{field_length}{field_length}
\item{data_type}{data_type}
\item{description}{description}
\item{comments}{comments}
\item{valid_values}{valid_values}
\item{spanner1}{spanner1}
\item{spanner2}{spanner2}
\item{final_name}{final_name}
#' }
}
\source{
NJDOE website - eg the 'File Layout' link on
http://www.state.nj.us/education/schools/achievement/2009/njask3/
}
\usage{
layout_njask09
}
\description{
A processed R version of the 'File Layout' for the 2009 NJASK   
Original file is in .xls format on NJDOE website
}
\keyword{datasets}
