% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\docType{data}
\name{LEGACY_ASSESS_VALID_YEARS}
\alias{LEGACY_ASSESS_VALID_YEARS}
\title{Legacy assessment (NJASK/HSPA/GEPA) valid years}
\format{
An object of class \code{integer} of length 11.
}
\usage{
LEGACY_ASSESS_VALID_YEARS
}
\description{
Legacy assessment (NJASK/HSPA/GEPA) valid years
}
\keyword{internal}
