% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_percentiles.R
\name{lookup_peer_percentile}
\alias{lookup_peer_percentile}
\title{Looks up caclulated percentile value by searching for closest
scale score / proficient above match}
\usage{
lookup_peer_percentile(assess_agg, assess_percentiles)
}
\arguments{
\item{assess_agg}{an assessments aggregate such as the output 
of \code{charter_sector_parcc_aggs}}

\item{assess_percentiles}{calculated assessments peer percentiles 
such as the output of \code{statewide_peer_percentile}}
}
\value{
data.frame with percent proficient and scale score percentile ranks
}
\description{
Given a peer percentile lookup table with calculated
mean scale score and percent proficient distributions by year, grade,
subgroup, take the given values (likely of assessment aggregates) and
find the closest match to return percentiles
}
