% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_cache_key}
\alias{make_cache_key}
\title{Generate a cache key for a data request}
\usage{
make_cache_key(fn_name, ...)
}
\arguments{
\item{fn_name}{Name of the fetch function}

\item{...}{Arguments passed to the fetch function}
}
\value{
Character string cache key
}
\description{
Generate a cache key for a data request
}
\keyword{internal}
