% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_calcs.R
\name{matric_aggregate_calcs}
\alias{matric_aggregate_calcs}
\title{Aggregate multiple postsecondary matriculation rows and produce
summary statistics}
\usage{
matric_aggregate_calcs(df)
}
\arguments{
\item{df}{grouped df of postsecondary matriculation data}
}
\value{
data_frame
}
\description{
Aggregate multiple postsecondary matriculation rows and produce
summary statistics
}
