% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{matric_column_order}
\alias{matric_column_order}
\title{Matriculation column order}
\usage{
matric_column_order(df)
}
\arguments{
\item{df}{processed postsecondary matriculation df}
}
\value{
df in correct order
}
\description{
Matriculation column order
}
