% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{MIDDLE_GRADES}
\alias{MIDDLE_GRADES}
\title{Middle school grades (6-8)}
\format{
An object of class \code{character} of length 3.
}
\usage{
MIDDLE_GRADES
}
\description{
Middle school grades (6-8)
}
\keyword{internal}
