% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_peer_groups.R
\docType{data}
\name{NEWARK_DISTRICT_ID}
\alias{NEWARK_DISTRICT_ID}
\title{Newark district ID constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
NEWARK_DISTRICT_ID
}
\description{
The district ID for Newark Public Schools (3570).
Used as a constant for filtering and analysis.
}
\keyword{datasets}
