% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nj_assess.R
\name{nj_coltype_parser}
\alias{nj_coltype_parser}
\title{nj_coltype_parser}
\usage{
nj_coltype_parser(datatypes)
}
\arguments{
\item{datatypes}{vector of datatypes (from a layout df)}
}
\value{
a character string of the types, for read_fwf
}
\description{
turns layout datatypes into compact string required by read_fwf
}
\keyword{internal}
