% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_config.R
\docType{data}
\name{njdoe_base_urls}
\alias{njdoe_base_urls}
\title{NJ DOE base URLs}
\format{
An object of class \code{list} of length 4.
}
\usage{
njdoe_base_urls
}
\description{
NJ DOE base URLs
}
\keyword{internal}
