% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njschooldata-package.R
\docType{package}
\name{njschooldata-package}
\alias{njschooldata}
\alias{njschooldata-package}
\title{njschooldata: A Simple Interface for Accessing NJ DOE School Data in R}
\description{
R functions to import NJ school data, including assessment, enrollment, and accountability data from the New Jersey Department of Education.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almartin82/njschooldata}
  \item Report bugs at \url{https://github.com/almartin82/njschooldata/issues}
}

}
\author{
\strong{Maintainer}: Andrew Martin \email{almartin@gmail.com}

}
\keyword{internal}
