% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{njsd_cache_enabled}
\alias{njsd_cache_enabled}
\title{Check if caching is enabled}
\usage{
njsd_cache_enabled()
}
\value{
Logical indicating whether caching is currently enabled
}
\description{
Check if caching is enabled
}
