% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwk_address_addendum}
\alias{nwk_address_addendum}
\title{Newark Address Addendum}
\format{
A data frame with 98 rows and 5 columns:
\describe{
  \item{district_id}{District identifier (Newark district is 3570)}
  \item{school_id}{School identifier code}
  \item{school_name}{Name of the school}
  \item{address}{Full street address with city, state}
  \item{in_geocode}{Logical indicating if address is in geocode cache}
}
}
\source{
Manual address corrections for Newark schools
}
\usage{
nwk_address_addendum
}
\description{
Additional address data for Newark schools to supplement geocoding.
Contains school addresses that were not included in the main school
directory or needed corrections for proper geocoding.
}
\keyword{datasets}
