% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pad_cds}
\alias{pad_cds}
\title{pad cds fields}
\usage{
pad_cds(df)
}
\arguments{
\item{df}{containing county_code, district_code, school_code}
}
\value{
data frame with zero padded cds columns
}
\description{
pad cds fields
}
