% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_padding.R
\name{pad_grade}
\alias{pad_grade}
\title{Pad grade level}
\usage{
pad_grade(x)
}
\arguments{
\item{x}{a grade level argument, length 1}
}
\value{
a string, length 2, with appropriate padding for PARCC naming conventions
}
\description{
Ensures grade level is two characters with leading zero if needed.
}
