% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc.R
\name{parcc_column_order}
\alias{parcc_column_order}
\title{PARCC column order}
\usage{
parcc_column_order(df)
}
\arguments{
\item{df}{tidied PARCC dataframe.  called as final step in fetch_parcc when tidy=TRUE}
}
\value{
PARCC df with columns in coherent order
}
\description{
PARCC column order
}
