% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{PARCC_LEVELS}
\alias{PARCC_LEVELS}
\title{PARCC/NJSLA performance levels}
\format{
An object of class \code{character} of length 5.
}
\usage{
PARCC_LEVELS
}
\description{
PARCC/NJSLA performance levels
}
\keyword{internal}
