% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\docType{data}
\name{PARCC_VALID_YEARS}
\alias{PARCC_VALID_YEARS}
\title{PARCC/NJSLA assessment valid years (skip 2020 - no testing due to COVID)}
\format{
An object of class \code{integer} of length 9.
}
\usage{
PARCC_VALID_YEARS
}
\description{
PARCC/NJSLA assessment valid years (skip 2020 - no testing due to COVID)
}
\keyword{internal}
