% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_parcc_subj}
\alias{parse_parcc_subj}
\title{Parse PARCC subject}
\usage{
parse_parcc_subj(x)
}
\arguments{
\item{x}{a subject (function parameter subj)}
}
\value{
the subject coded as ELA or MAT.
}
\description{
Parse PARCC subject
}
