% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{peek}
\alias{peek}
\title{Peek at a data frame}
\usage{
peek(df, nrows = 5)
}
\arguments{
\item{df}{data.frame}

\item{nrows}{how many rows to sample}
}
\value{
prints random sample of nrows of a dataframe
}
\description{
Displays a random sample of rows from a data frame for quick inspection.
}
\keyword{internal}
