% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{percentile_rank}
\alias{percentile_rank}
\title{Percentile Rank}
\usage{
percentile_rank(x, xo)
}
\arguments{
\item{x}{vector of values}

\item{xo}{target value}
}
\value{
numeric percentile rank
}
\description{
Percentile Rank
}
