% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{PREK_PROGRAM_CODES}
\alias{PREK_PROGRAM_CODES}
\title{Pre-K program codes}
\format{
An object of class \code{character} of length 4.
}
\usage{
PREK_PROGRAM_CODES
}
\description{
Pre-K program codes
}
\keyword{internal}
