% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{print.njsd_cache_info}
\alias{print.njsd_cache_info}
\title{Print cache information}
\usage{
\method{print}{njsd_cache_info}(x, ...)
}
\arguments{
\item{x}{Object from njsd_cache_info()}

\item{...}{Additional arguments (unused)}
}
\description{
Print cache information
}
