% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{process_access}
\alias{process_access}
\title{Process raw ACCESS for ELLs data}
\usage{
process_access(access_file, end_year)
}
\arguments{
\item{access_file}{Output of get_raw_access}

\item{end_year}{A school year (2022-2024)}
}
\value{
Processed ACCESS dataframe
}
\description{
Cleans and standardizes ACCESS data.
}
\keyword{internal}
