% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{process_days_absent_cols}
\alias{process_days_absent_cols}
\title{Process days absent columns}
\usage{
process_days_absent_cols(df)
}
\arguments{
\item{df}{Raw data frame from SPR database}
}
\value{
Data frame with standardized columns
}
\description{
Standardizes column names for days absent data.
}
\keyword{internal}
