% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr}
\alias{process_enr}
\title{Process a NJ enrollment file}
\usage{
process_enr(df)
}
\arguments{
\item{df}{An enrollment data frame (eg output of `get_raw_enr`)}
}
\value{
Cleaned and processed enrollment data frame
}
\description{
Does cleanup of dataframes returned by `get_raw_enr`.
}
\keyword{internal}
