% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enr.R
\name{process_enr_program}
\alias{process_enr_program}
\title{join program code to program name}
\usage{
process_enr_program(df)
}
\arguments{
\item{df}{cleaned enrollment file}
}
\description{
decode the program name
}

