% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{process_grad_count}
\alias{process_grad_count}
\title{Process Grad Count Data}
\usage{
process_grad_count(df, end_year)
}
\arguments{
\item{df}{output of get_grad_count}

\item{end_year}{end of the academic year}
}
\value{
dataframe with composite subgroups like black (black_m + black_f)
}
\description{
Process Grad Count Data
}
