% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\name{process_grate}
\alias{process_grate}
\title{Process graduation rate data}
\usage{
process_grate(df, end_year)
}
\arguments{
\item{df}{The output of get_raw_grad_file}

\item{end_year}{A school year. Year is the end of the academic year - eg 2006-07
school year is year '2007'. Valid values are 1998-2024.}
}
\value{
Cleaned graduation data frame
}
\description{
Does cleanup of the grad rate ('grate') file.
}
\keyword{internal}
