% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_nj_assess}
\alias{process_nj_assess}
\title{Process a NJ assessment file}
\usage{
process_nj_assess(df, layout)
}
\arguments{
\item{df}{A raw NJASK, HSPA, or GEPA data frame (eg output of `get_raw_njask`)}

\item{layout}{Which layout file to use to determine which columns are one implied
decimal.}
}
\value{
Processed assessment data frame
}
\description{
Does cleanup of the raw assessment file, primarily ensuring that
columns tagged as 'one implied' are displayed correctly.
}
\keyword{internal}
