% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/njask.R
\name{process_njask}
\alias{process_njask}
\title{process a raw njask file}
\usage{
process_njask(df, mask = layout_njask$comments)
}
\arguments{
\item{df}{a NJASK data frame (output of \code{get_raw_njask})
school year is year '2014'.  valid values are 2004-2014.}

\item{mask}{a vector indicating which columns are one implied decimal?
uses a layout file.  default is layout_njask}
}
\description{
\code{process_njask} does cleanup of the raw njask file, primarily ensuring that
columns tagged as 'one implied' are displayed correctly#'
}

