% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{process_njdoe_cds_facilities}
\alias{process_njdoe_cds_facilities}
\title{Process NJDOE CDS workbook to inventory and closure lifecycle rows}
\usage{
process_njdoe_cds_facilities(raw, src)
}
\description{
Process NJDOE CDS workbook to inventory and closure lifecycle rows
}
\keyword{internal}
