% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{process_njdoe_lead_soa}
\alias{process_njdoe_lead_soa}
\title{Process NJDOE lead Statement of Assurance workbook}
\usage{
process_njdoe_lead_soa(raw, src)
}
\description{
Process NJDOE lead Statement of Assurance workbook
}
\keyword{internal}
