% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{process_njdoe_sda_allocation}
\alias{process_njdoe_sda_allocation}
\title{Process NJDOE SDA district allocation workbook}
\usage{
process_njdoe_sda_allocation(raw, src)
}
\description{
Process NJDOE SDA district allocation workbook
}
\keyword{internal}
