% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{process_njgin_school_points}
\alias{process_njgin_school_points}
\title{Process NJGIN school points to inventory rows}
\usage{
process_njgin_school_points(raw, src)
}
\description{
Process NJGIN school points to inventory rows
}
\keyword{internal}
