% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_facilities.R
\name{process_njsda_active_projects}
\alias{process_njsda_active_projects}
\title{Process NJSDA active project detail pages}
\usage{
process_njsda_active_projects(raw, src)
}
\description{
Process NJSDA active project detail pages
}
\keyword{internal}
