% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc.R
\name{process_parcc}
\alias{process_parcc}
\title{Process a raw PARCC data file}
\usage{
process_parcc(parcc_file, end_year, grade, subj)
}
\arguments{
\item{parcc_file}{output of get_raw_parcc}

\item{end_year}{a school year.  end_year is the end of the academic year - eg 2014-15
school year is end_year '2015'.  valid values are 2015}

\item{grade}{grade level}

\item{subj}{PARCC subject. c('ela' or 'math')}
}
\value{
a tbl_df / data rame
}
\description{
all the logic needed to clean up the raw PARCC files
}

