% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_pop.R
\name{process_reportcard_special_pop}
\alias{process_reportcard_special_pop}
\title{Process Report Card Special Population Data}
\usage{
process_reportcard_special_pop(df)
}
\arguments{
\item{df}{raw special pop dataframe - output of get_reportcard_special_pop}
}
\value{
tidy dataframe with conforming id columns
}
\description{
Process Report Card Special Population Data
}
