% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{PROFICIENT_LEVELS}
\alias{PROFICIENT_LEVELS}
\title{Proficient levels (L4 and L5)}
\format{
An object of class \code{character} of length 2.
}
\usage{
PROFICIENT_LEVELS
}
\description{
Proficient levels (L4 and L5)
}
\keyword{internal}
