% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/other_data.R
\docType{data}
\name{prog_codes}
\alias{prog_codes}
\title{NJ Program Codes}
\format{
A data frame with 682 rows and 3 columns:
\describe{
  \item{end_year}{School year ending year}
  \item{program_code}{Two-digit program code}
  \item{program_name}{Program name or description}
}

\describe{
\item{end_year}{end_year}
\item{program_code}{program_code}
\item{program_name}{program_name}
#' }
}
\source{
NJ Department of Education

NJDOE website for each year of enrollment data (see link above)
}
\usage{
prog_codes

prog_codes
}
\description{
Program codes used by the New Jersey Department of Education for student enrollment
reporting. Includes grade levels, special education programs, and other educational
programs.

A compilation of all the "program codes" used by the state of NJ. example: 
http://www.nj.gov/education/data/enr/enr15/stat_doc.htm
}
\keyword{datasets}
