% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_enabled}
\alias{progress_enabled}
\title{Check if progress indicators are enabled}
\usage{
progress_enabled()
}
\value{
Logical
}
\description{
Check if progress indicators are enabled
}
\keyword{internal}
