% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_tracker}
\alias{progress_tracker}
\title{Create a simple progress tracker}
\usage{
progress_tracker(total, task_name = "Processing")
}
\arguments{
\item{total}{Total number of items to process}

\item{task_name}{Name of the task being performed}
}
\value{
A list with update() and done() functions
}
\description{
Creates a progress tracker for batch operations that displays
progress messages to the console.
}
\examples{
\dontrun{
pb <- progress_tracker(10, "Downloading files")
for (i in 1:10) {
  Sys.sleep(0.1)
  pb$update(i, sprintf("File \%d", i))
}
pb$done()
}
}
\keyword{internal}
