% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rc_numeric_cleaner}
\alias{rc_numeric_cleaner}
\title{Report Card Numeric Data Cleaner}
\usage{
rc_numeric_cleaner(data_vector)
}
\arguments{
\item{data_vector}{vector of data that has percent signs, supression codes, or 
a variety of representations of N/A}
}
\value{
numeric vector
}
\description{
Report Card Numeric Data Cleaner
}
