% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rc_year_matcher}
\alias{rc_year_matcher}
\title{report card year matcher}
\usage{
rc_year_matcher(df)
}
\arguments{
\item{df}{a report card table that includes trailing/longitudinal data}
}
\value{
data frame with only data for the year of the report
}
\description{
report card year matcher
}
