% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{school_name_to_id}
\alias{school_name_to_id}
\title{Given friendly school names, returns ids}
\usage{
school_name_to_id(school_names, df)
}
\arguments{
\item{school_names}{vector or list of friendly school names}

\item{df}{df used to generate the friendly names}
}
\value{
vector of school_ids
}
\description{
Given friendly school names, returns ids
}
