% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_pop_calcs.R
\name{spec_pop_aggregate_calcs}
\alias{spec_pop_aggregate_calcs}
\title{Aggregate multiple special populations rows and produce summary
statistics}
\usage{
spec_pop_aggregate_calcs(df)
}
\arguments{
\item{df}{grouped df of special populations data}
}
\value{
data_frame
}
\description{
Aggregate multiple special populations rows and produce summary
statistics
}
