% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{split_enr_cols}
\alias{split_enr_cols}
\title{Split enrollment columns}
\usage{
split_enr_cols(df)
}
\arguments{
\item{df}{An enrollment data frame (eg output of `get_raw_enr`)}
}
\value{
Data frame with split columns
}
\description{
Splits enrollment columns that combine IDs and names (pre 2009-10 format)
}
\keyword{internal}
