% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_percentiles.R
\name{statewide_peer_percentile}
\alias{statewide_peer_percentile}
\title{Calculate statewide peer percentile by grade}
\usage{
statewide_peer_percentile(df)
}
\arguments{
\item{df}{}
}
\value{
data.frame with percent proficient and scale score percentile rank
}
\description{
calculates statewide percentile by grade/test
}
