% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tges_name_cleaner}
\alias{tges_name_cleaner}
\title{TGES name cleaner}
\usage{
tges_name_cleaner(x, indicator_fields)
}
\arguments{
\item{x}{vector of names}

\item{indicator_fields}{list of key/value variables to convert}
}
\value{
character vector of names
}
\description{
internal function for converting cryptic variable codes to full name
}
