% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_admin_salaries}
\alias{tidy_admin_salaries}
\title{Tidy Administrative Salaries and Benefits}
\usage{
tidy_admin_salaries(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Administrative Salaries and Benefits
}
