% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_administrative_costs}
\alias{tidy_administrative_costs}
\title{Tidy Administrative Costs}
\usage{
tidy_administrative_costs(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Administrative Costs
}
\keyword{internal}
