% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_budgetary_per_pupil_cost}
\alias{tidy_budgetary_per_pupil_cost}
\title{Tidy Budgetary Per Pupil data frame}
\usage{
tidy_budgetary_per_pupil_cost(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Budgetary Per Pupil data frame
}
