% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_budgeted_vs_actual_fund_balance}
\alias{tidy_budgeted_vs_actual_fund_balance}
\title{tidy budgeted vs actual fund balance}
\usage{
tidy_budgeted_vs_actual_fund_balance(df, end_year)
}
\arguments{
\item{df}{general fund vs actual used data frame, eg CSG20 
output from get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data frame
}
\description{
tidy budgeted vs actual fund balance
}
