% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_classroom_general_supplies}
\alias{tidy_classroom_general_supplies}
\title{Tidy Classroom General Supplies and Textbooks}
\usage{
tidy_classroom_general_supplies(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Classroom General Supplies and Textbooks
}
\keyword{internal}
