% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_classroom_purchased_services}
\alias{tidy_classroom_purchased_services}
\title{Tidy Classroom Purchased Services and Other}
\usage{
tidy_classroom_purchased_services(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Classroom Purchased Services and Other
}
\keyword{internal}
