% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_classroom_salaries_benefits}
\alias{tidy_classroom_salaries_benefits}
\title{Tidy Classroom Salaries and Benefits}
\usage{
tidy_classroom_salaries_benefits(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Classroom Salaries and Benefits
}
\keyword{internal}
