% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_excess_unreserved_general_fund}
\alias{tidy_excess_unreserved_general_fund}
\title{Tidy Excess Unreserved General Fund}
\usage{
tidy_excess_unreserved_general_fund(df, end_year)
}
\arguments{
\item{df}{excess unreserved general fund data frame, eg CSG21 
output from get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data frame
}
\description{
Tidy Excess Unreserved General Fund
}
