% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_food_service}
\alias{tidy_food_service}
\title{Tidy Food Service Costs}
\usage{
tidy_food_service(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Food Service Costs
}
\keyword{internal}
