% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_generic_budget_indicator}
\alias{tidy_generic_budget_indicator}
\title{tidy common/generic budget indicator data frame}
\usage{
tidy_generic_budget_indicator(df, end_year, indicator)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges() 
indicators 1-15}

\item{end_year}{end year that the report was published}

\item{indicator}{character, indicator name}
}
\value{
long, tidy data frame
}
\description{
tidy common/generic budget indicator data frame
}
