% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_generic_personnel}
\alias{tidy_generic_personnel}
\title{tidy generic personnel indicator data frame}
\usage{
tidy_generic_personnel(df, end_year, indicator)
}
\arguments{
\item{df}{personnel data frame, eg output of get_raw_tges() 
indicators 16-19}

\item{end_year}{end year that the report was published}

\item{indicator}{character, indicator name}
}
\value{
long, tidy data frame
}
\description{
tidy generic personnel indicator data frame
}
