% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_graduation.R
\name{tidy_grad_count}
\alias{tidy_grad_count}
\title{Tidy Grad Count}
\usage{
tidy_grad_count(df, end_year)
}
\arguments{
\item{df}{Output of process_grad_count}

\item{end_year}{End of the academic year - eg 2006-07 is 2007.
Valid values are 1998-present.}
}
\value{
data.frame with number of graduates
}
\description{
Transforms graduation count data to long format.
}
\keyword{internal}
