% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_graduation.R
\name{tidy_grad_rate}
\alias{tidy_grad_rate}
\title{Tidy grad rate}
\usage{
tidy_grad_rate(df, end_year, methodology = "4 year")
}
\arguments{
\item{df}{The output of process_grad_rate}

\item{end_year}{A school year. Year is the end of the academic year - eg 2006-07
school year is year '2007'. Valid values are 1998-2024.}

\item{methodology}{One of '4 year' or '5 year'}
}
\value{
Tidied graduation rate data frame
}
\description{
Tidies a processed grate data frame, producing a data frame with consistent
headers and values, suitable for longitudinal analysis.
}
\keyword{internal}
