% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{tidy_grate}
\alias{tidy_grate}
\title{tidy grate}
\usage{
tidy_grate(df, end_year)
}
\arguments{
\item{df}{the output of process_grate}

\item{end_year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 1998-2018.}
}
\description{
tidies a processed grate data frame, producing a data frame with consistent
headers and values, suitable for longitudinal analysis
}
