% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_legal_services}
\alias{tidy_legal_services}
\title{Tidy Legal Services}
\usage{
tidy_legal_services(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Legal Services
}
