% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{tidy_parcc_subgroup}
\alias{tidy_parcc_subgroup}
\title{Tidy PARCC subgroup names}
\usage{
tidy_parcc_subgroup(sv)
}
\arguments{
\item{sv}{Subgroup column from PARCC data file}
}
\value{
Character vector with consistent subgroup names
}
\description{
Standardizes subgroup names across years.
}
