% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_personal_services_benefits}
\alias{tidy_personal_services_benefits}
\title{Tidy Personal Services and Benefits Costs}
\usage{
tidy_personal_services_benefits(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Personal Services and Benefits Costs
}
