% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_plant_operations_maintenance}
\alias{tidy_plant_operations_maintenance}
\title{Tidy Plant Operations and Maintenance}
\usage{
tidy_plant_operations_maintenance(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Plant Operations and Maintenance
}
