% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_ratio_faculty_to_administrators}
\alias{tidy_ratio_faculty_to_administrators}
\title{Tidy Ratio of Faculty to Administrators}
\usage{
tidy_ratio_faculty_to_administrators(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Ratio of Faculty to Administrators
}
