% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_ratio_students_to_special_service}
\alias{tidy_ratio_students_to_special_service}
\title{Tidy Ratio of Students to Special Service}
\usage{
tidy_ratio_students_to_special_service(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Ratio of Students to Special Service
}
\keyword{internal}
