% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{tidy_sped_placement_state}
\alias{tidy_sped_placement_state}
\title{Tidy the State by Ed Environ sheet (5-21 or 3-5) -- 2025}
\usage{
tidy_sped_placement_state(df, age_group)
}
\description{
Tidy the State by Ed Environ sheet (5-21 or 3-5) -- 2025
}
\keyword{internal}
