% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_support_services_salaries}
\alias{tidy_support_services_salaries}
\title{Tidy Support Services Salaries}
\usage{
tidy_support_services_salaries(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Support Services Salaries
}
