% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_tges_data}
\alias{tidy_tges_data}
\title{Tidy list of TGES data frames}
\usage{
tidy_tges_data(list_of_dfs, end_year)
}
\arguments{
\item{list_of_dfs}{list of TGES data frames, eg output of 
get_raw_tges(). Current valid values are 2011 to 2017.}

\item{end_year}{year that the report was published}
}
\value{
list of cleaned (wide to long, tidy) dataframes
}
\description{
Tidy list of TGES data frames
}
