% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_total_classroom_instruction}
\alias{tidy_total_classroom_instruction}
\title{Tidy Total Classroom Instruction data frame}
\usage{
tidy_total_classroom_instruction(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Total Classroom Instruction data frame
}
\keyword{internal}
