% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_total_spending_per_pupil}
\alias{tidy_total_spending_per_pupil}
\title{tidy total spending per pupil}
\usage{
tidy_total_spending_per_pupil(df, end_year)
}
\arguments{
\item{df}{total spending data frame, eg CSG1AA_AVGS output from get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data frame
}
\description{
tidy total spending per pupil
}
\keyword{internal}
