% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_total_support_services}
\alias{tidy_total_support_services}
\title{Tidy Total Support Services}
\usage{
tidy_total_support_services(df, end_year)
}
\arguments{
\item{df}{indicator data frame, eg output of get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data.frame
}
\description{
Tidy Total Support Services
}
