% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tidy_vitstat}
\alias{tidy_vitstat}
\title{Tidy Vital Statistics}
\usage{
tidy_vitstat(df, end_year)
}
\arguments{
\item{df}{vital statistics data frame, eg VITSTAT_TOTAL 
output from get_raw_tges()}

\item{end_year}{end year that the report was published}
}
\value{
data frame
}
\description{
Tidy Vital Statistics
}
