% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{TOTAL_PROGRAM_CODE}
\alias{TOTAL_PROGRAM_CODE}
\title{Program code for school/district totals}
\format{
An object of class \code{character} of length 1.
}
\usage{
TOTAL_PROGRAM_CODE
}
\description{
Program code for school/district totals
}
\keyword{internal}
