% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Trim whitespace}
\usage{
trim_whitespace(x)
}
\arguments{
\item{x}{string or vector of strings}
}
\value{
a string or vector of strings, with whitespace removed.
}
\description{
Trim whitespace
}

