% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trunc2}
\alias{trunc2}
\title{Truncate with configurable precision}
\usage{
trunc2(x, prec = 0)
}
\arguments{
\item{x}{numeric vector}

\item{prec}{desired precision}
}
\value{
truncated numeric vector
}
\description{
Truncates numeric values to specified decimal precision.
}
