% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unzipper}
\alias{unzipper}
\title{Wrapper to download a file from a url and unzip}
\usage{
unzipper(url, file_pattern = "njschooldata")
}
\arguments{
\item{url}{character, target url}

\item{file_pattern}{charcter, stub to use for temporary file}
}
\value{
list of unzipped files in a temp directory
}
\description{
Wrapper to download a file from a url and unzip
}
