% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_logical}
\alias{validate_logical}
\title{Validate logical/boolean parameter}
\usage{
validate_logical(value, param_name)
}
\arguments{
\item{value}{The value to validate}

\item{param_name}{The name of the parameter (for error messages)}
}
\value{
TRUE invisibly if valid, otherwise throws an error
}
\description{
Validate logical/boolean parameter
}
