% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_parcc_call}
\alias{validate_parcc_call}
\title{Validate PARCC/NJSLA parameters}
\usage{
validate_parcc_call(end_year, grade_or_subj, subj)
}
\arguments{
\item{end_year}{The year}

\item{grade_or_subj}{The grade or subject code}

\item{subj}{The subject ("ela" or "math")}
}
\value{
TRUE invisibly if valid, otherwise throws an error
}
\description{
Validates all parameters for fetch_parcc() in one call.
}
