% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ward_gcount_aggs}
\alias{ward_gcount_aggs}
\title{Aggregates grad counts data by ward}
\usage{
ward_gcount_aggs(df)
}
\arguments{
\item{df}{output of \code{fetch_grad_count}}
}
\value{
A data frame of ward aggregations
}
\description{
Aggregates grad counts data by ward
}
