% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ward_grate_aggs}
\alias{ward_grate_aggs}
\title{Aggregates grad rate data by ward}
\usage{
ward_grate_aggs(df)
}
\arguments{
\item{df}{output of \code{fetch_grad_rate}}
}
\value{
A data frame of ward aggregations
}
\description{
Aggregates grad rate data by ward
}
