% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ward_matric_aggs}
\alias{ward_matric_aggs}
\title{Aggregates matriculation data by ward}
\usage{
ward_matric_aggs(df)
}
\arguments{
\item{df}{output of \code{enrich_matric_counts}}
}
\value{
A data frame of ward aggregations
}
\description{
Aggregates matriculation data by ward
}
