% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{WARD_SUFFIX}
\alias{WARD_SUFFIX}
\title{Suffix for ward aggregations}
\format{
An object of class \code{character} of length 1.
}
\usage{
WARD_SUFFIX
}
\description{
Suffix for ward aggregations
}
\keyword{internal}
