% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{year_variable_converter}
\alias{year_variable_converter}
\title{year variable converter}
\usage{
year_variable_converter(df, end_year)
}
\arguments{
\item{df}{a tges indicator data frame published between 1999 and 2003}

\item{end_year}{year published}
}
\value{
data frame that conforms to 2004-2009 style
}
\description{
for the 1999-2003 tges files, the 'year' of the data
was encoded in the variable names
}
