% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{allpublic_gcount_aggs}
\alias{allpublic_gcount_aggs}
\title{Calculate Charter Sector Grad Count aggregates}
\usage{
allpublic_gcount_aggs(df)
}
\arguments{
\item{df}{tidied grad count dataframe, eg output of fetch_grad_count}
}
\value{
df containing charter sector aggregate grad count
}
\description{
Calculate Charter Sector Grad Count aggregates
}
