% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{ALT_SCHOOL_ID}
\alias{ALT_SCHOOL_ID}
\title{Alternative school code (used in some graduation files)}
\format{
An object of class \code{character} of length 1.
}
\usage{
ALT_SCHOOL_ID
}
\description{
Alternative school code (used in some graduation files)
}
\keyword{internal}
