% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{build_sped_placement_url}
\alias{build_sped_placement_url}
\title{Build the IDEA 618 placement workbook URL}
\usage{
build_sped_placement_url(end_year)
}
\arguments{
\item{end_year}{ending school year (currently only 2025)}
}
\value{
character URL
}
\description{
Build the IDEA 618 placement workbook URL
}
\keyword{internal}
