% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{CHARTER_COUNTY_ID}
\alias{CHARTER_COUNTY_ID}
\title{Charter county code (all charters are assigned to county 80)}
\format{
An object of class \code{character} of length 1.
}
\usage{
CHARTER_COUNTY_ID
}
\description{
Charter county code (all charters are assigned to county 80)
}
\keyword{internal}
