% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_cds_fields}
\alias{clean_cds_fields}
\title{Clean up CDS field names}
\usage{
clean_cds_fields(df)
}
\arguments{
\item{df}{}
}
\value{
df, with consistent county_code, district_code, school_code fields
}
\description{
Clean up CDS field names
}
