% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{clean_enr_data}
\alias{clean_enr_data}
\title{clean enrollment data types}
\usage{
clean_enr_data(df)
}
\arguments{
\item{df}{a enr data frame (eg output of \code{get_raw_enr})}
}
\description{
all columns come back char; coerce some back to numeric
}

