% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_names.R
\name{clean_enr_names}
\alias{clean_enr_names}
\title{Clean enrollment column names}
\usage{
clean_enr_names(df)
}
\arguments{
\item{df}{An enrollment data frame (eg output of `get_raw_enr`)}
}
\value{
Data frame with standardized column names
}
\description{
Give consistent names to the enrollment files across all years.
}
\keyword{internal}
