% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped.R
\name{clean_sped_df}
\alias{clean_sped_df}
\title{Clean SPED data}
\usage{
clean_sped_df(df, end_year)
}
\arguments{
\item{df}{raw data frame with cleaned names, output of get_raw_sped with clean_sped_names applied.}

\item{end_year}{academic year, ending year - eg 2023-2024 is 2024.}
}
\value{
cleaned data frame
}
\description{
Cleans and standardizes SPED data from NJ DOE.
}
