% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nj_assess.R
\name{common_fwf_req}
\alias{common_fwf_req}
\title{common_fwf_req}
\usage{
common_fwf_req(url, layout)
}
\arguments{
\item{url}{file location}

\item{layout}{data frame containing fixed-width file column specifications}
}
\value{
layout layout to use
}
\description{
common fwf logic across various assessment types.  DRY.
}
