% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{DISTRICT_TOTAL_SCHOOL_ID}
\alias{DISTRICT_TOTAL_SCHOOL_ID}
\title{District total school code}
\format{
An object of class \code{character} of length 1.
}
\usage{
DISTRICT_TOTAL_SCHOOL_ID
}
\description{
District total school code
}
\keyword{internal}
