% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{download_and_clean_pr}
\alias{download_and_clean_pr}
\title{Download and clean performance report data}
\usage{
download_and_clean_pr(tmp_pr, url, end_year)
}
\arguments{
\item{tmp_pr}{path to tempfile}

\item{url}{url to download}

\item{end_year}{report year}
}
\value{
list of dataframes
}
\description{
Download and clean performance report data
}
