% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{enr_grade_aggs}
\alias{enr_grade_aggs}
\title{Custom Enrollment Grade Level Aggregates}
\usage{
enr_grade_aggs(df)
}
\arguments{
\item{df}{a tidy enrollment df}
}
\value{
df of aggregated enrollment data
}
\description{
Custom Enrollment Grade Level Aggregates
}
