% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\docType{data}
\name{ENR_VALID_YEARS}
\alias{ENR_VALID_YEARS}
\title{Enrollment data valid years
Note: 1999 data was removed from NJ DOE website}
\format{
An object of class \code{integer} of length 26.
}
\usage{
ENR_VALID_YEARS
}
\description{
Enrollment data valid years
Note: 1999 data was removed from NJ DOE website
}
\keyword{internal}
