% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{enrich_grad_count}
\alias{enrich_grad_count}
\title{Enrich report card matriculation percentages with best guesses at
graduated students}
\usage{
enrich_grad_count(df, end_year)
}
\arguments{
\item{df}{data frame of including subgroup percentages}

\item{end_year}{numeric end year of grad counts to join}
}
\value{
data_frame
}
\description{
Enrich report card matriculation percentages with best guesses at
graduated students
}
