% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{enrich_rc_enrollment}
\alias{enrich_rc_enrollment}
\title{Enrich report card subgroup percentages with best guesses at 
subgroup numbers}
\usage{
enrich_rc_enrollment(df)
}
\arguments{
\item{df}{data frame of icluding subgroup percentages

return data_frame}
}
\description{
Enrich report card subgroup percentages with best guesses at 
subgroup numbers
}
