% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{enrich_school_latlong}
\alias{enrich_school_latlong}
\title{Enrich School Data with Lat / Long}
\usage{
enrich_school_latlong(df, use_cache = TRUE, api_key = "")
}
\arguments{
\item{df}{dataframe to be enriched}

\item{use_cache}{if TRUE, will read from cache of school info / lat lng stored on TODO}

\item{api_key}{optional, personal google maps API key}
}
\value{
dataframe enriched with lat lng
}
\description{
Enrich School Data with Lat / Long
}
\note{
The `placement` package is required for geocoding when `use_cache=FALSE`.
  Install with: `remotes::install_github('DerekYves/placement')`
}
