% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{extract_rc_college_matric}
\alias{extract_rc_college_matric}
\title{Extract Report Card Matriculation Rates}
\usage{
extract_rc_college_matric(
  list_of_prs,
  school_only = TRUE,
  cds_identifiers = TRUE
)
}
\arguments{
\item{list_of_prs}{output of get_rc_databases (ie, a list where each element is)
a list of data.frames}

\item{school_only}{some years have district average results, not just school-level. 
if school_only, return only school data.  default is TRUE}

\item{cds_identifiers}{add the county, district and school name?  default is TRUE}
}
\value{
data frame with all the years of 4 year and 2 year college matriculation data
present in in the input
}
\description{
Extract Report Card Matriculation Rates
}
