% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_facilities.R
\name{facilities_cache_dir}
\alias{facilities_cache_dir}
\title{Internal facilities cache directory}
\usage{
facilities_cache_dir()
}
\description{
Internal facilities cache directory
}
\keyword{internal}
