% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{fetch_all_access}
\alias{fetch_all_access}
\title{Fetch all ACCESS for ELLs results}
\usage{
fetch_all_access()
}
\value{
A data frame with all ACCESS results (2022-2024, all grades)
}
\description{
Convenience function to download and combine all available ACCESS
for ELLs results into a single data frame.
}
\examples{
\dontrun{
# Get all ACCESS results (takes a while)
all_access <- fetch_all_access()
}
}
