% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{fetch_all_chronic_absenteeism}
\alias{fetch_all_chronic_absenteeism}
\title{Fetch all Chronic Absenteeism data}
\usage{
fetch_all_chronic_absenteeism()
}
\value{
A data frame with all chronic absenteeism results (2017-2019, 2022-2024)
}
\description{
Convenience function to download and combine all available chronic
absenteeism data into a single data frame.
}
\examples{
\dontrun{
# Get all chronic absenteeism data
all_ca <- fetch_all_chronic_absenteeism()
}
}
