% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{fetch_all_parcc}
\alias{fetch_all_parcc}
\title{Fetch all PARCC results}
\usage{
fetch_all_parcc()
}
\value{
A data frame with all PARCC results
}
\description{
Convenience function to download and combine all PARCC results
into single data frame.
}
\examples{
\dontrun{
# Get all PARCC/NJSLA results (takes a while)
all_parcc <- fetch_all_parcc()
}
}
