% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc.R
\name{fetch_all_parcc}
\alias{fetch_all_parcc}
\title{Fetch all PARCC results}
\usage{
fetch_all_parcc()
}
\value{
a data frame with all PARCC results
}
\description{
convenience function to download and combine all PARCC results
into single data frame
}
