% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfg.R
\name{fetch_dfg}
\alias{fetch_dfg}
\title{Fetch NJ District Factor Group (DFG) data}
\usage{
fetch_dfg(revision = 2000)
}
\arguments{
\item{revision}{c(2000, 1990)}
}
\value{
data.frame
}
\description{
Fetch NJ District Factor Group (DFG) data
}
