% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{fetch_enr}
\alias{fetch_enr}
\title{gets and processes a NJ enrollment file}
\usage{
fetch_enr(end_year)
}
\arguments{
\item{end_year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 1999-2015.}
}
\description{
\code{fetch_enr} is a wrapper around \code{get_raw_enr} and
\code{process_enr} that passes the correct file layout data to each function,
given an end_year
}
