% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_enrollment.R
\name{fetch_enr}
\alias{fetch_enr}
\title{Gets and processes a NJ enrollment file}
\usage{
fetch_enr(end_year, tidy = FALSE)
}
\arguments{
\item{end_year}{A school year. Year is the end of the academic year - eg 2006-07
school year is year '2007'. Valid values are 1999-2025.}

\item{tidy}{If TRUE, takes the unwieldy wide data and normalizes into a
long, tidy data frame with limited headers - constants (school/district name and code),
subgroup (all the enrollment file subgroups), program/grade and measure (row_total, free lunch, etc).}
}
\value{
Data frame with processed enrollment data
}
\description{
`fetch_enr` is a wrapper around `get_raw_enr` and `process_enr` that
downloads and cleans enrollment data for a given year.
}
\examples{
\dontrun{
# Get 2023 enrollment data
enr_2023 <- fetch_enr(2023)

# Get tidy (long format) enrollment data
enr_tidy <- fetch_enr(2023, tidy = TRUE)
}
}
