% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{fetch_enr_years}
\alias{fetch_enr_years}
\title{Fetch multiple years of enrollment data with progress}
\usage{
fetch_enr_years(years, tidy = TRUE)
}
\arguments{
\item{years}{Vector of years to fetch}

\item{tidy}{Return tidy format? (default TRUE)}
}
\value{
Data frame with all enrollment data
}
\description{
Fetch multiple years of enrollment data with progress
}
\examples{
\dontrun{
enr_5yr <- fetch_enr_years(2020:2024)
}
}
