% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{fetch_essa_status}
\alias{fetch_essa_status}
\title{Fetch ESSA Accountability Status}
\usage{
fetch_essa_status(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with ESSA status ratings
}
\description{
Downloads ESSA accountability ratings from SPR database.
}
\examples{
\dontrun{
essa <- fetch_essa_status(2024)
}
}
