% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gepa.R
\name{fetch_gepa}
\alias{fetch_gepa}
\title{gets and processes a GEPA file}
\usage{
fetch_gepa(year)
}
\arguments{
\item{year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 2004-2007.}
}
\description{
\code{fetch_gepa} is a wrapper around \code{get_raw_gepa} and
\code{process_nj_assess} that passes the correct file layout data to each function,
given a year and grade.
}

