% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{fetch_grad_count}
\alias{fetch_grad_count}
\title{Fetch Grad Counts}
\usage{
fetch_grad_count(end_year, tidy = TRUE)
}
\arguments{
\item{end_year}{end of the academic year - eg 2006-07 is 2007.
valid values are 1998-present.}

\item{tidy}{if TRUE (default) will convert data to long format, where 
each row is a subgroup}
}
\value{
dataframe with grad counts
}
\description{
Fetch Grad Counts
}
