% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_graduation.R
\name{fetch_grad_rate}
\alias{fetch_grad_rate}
\title{Fetch Grad Rate}
\usage{
fetch_grad_rate(end_year, methodology = "4 year")
}
\arguments{
\item{end_year}{End of the academic year - eg 2006-07 is 2007.
Valid values are 2011-2024.}

\item{methodology}{Character string specifying calculation methodology.
One of "4 year" or "5 year".}
}
\value{
dataframe with grad rate
}
\description{
Downloads and processes graduation rate data.
}
\examples{
\dontrun{
# Get 2023 graduation rates
grate_2023 <- fetch_grad_rate(2023)

# Get 5-year graduation rates
grate_5yr <- fetch_grad_rate(2019, methodology = "5 year")
}
}
