% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{fetch_grate}
\alias{fetch_grate}
\title{fetch grate}
\usage{
fetch_grate(end_year, calc_type = "4 year", tidy = TRUE)
}
\arguments{
\item{end_year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 1998-2018.}

\item{calc_type}{c('4 year', '5 year')}

\item{tidy}{tidy the response into a data frame with consistent headers?}
}
\value{
graduation rate data frame
}
\description{
a consistent interface into the NJ HS graduates data.
}
