% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{fetch_many_tges}
\alias{fetch_many_tges}
\title{Fetch Multiple Cleaned Taxpayer's Guides to Educational Spending}
\usage{
fetch_many_tges(end_year_vector)
}
\arguments{
\item{end_year_vector}{vector of years.  Current valid values 
are 2011 to 2017.}
}
\value{
list of lists of data frames
}
\description{
Fetch Multiple Cleaned Taxpayer's Guides to Educational Spending
}
