% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{fetch_math_course_enrollment}
\alias{fetch_math_course_enrollment}
\title{Fetch Math Course Enrollment Data}
\usage{
fetch_math_course_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with math course enrollment
}
\description{
Downloads math course participation data from SPR database.
}
\examples{
\dontrun{
math <- fetch_math_course_enrollment(2024)
}
}
