% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgp.R
\name{fetch_msgp}
\alias{fetch_msgp}
\title{Fetch mSGP}
\usage{
fetch_msgp(end_year)
}
\arguments{
\item{end_year}{ending school year.  valid values are currently 2012-2018.}
}
\value{
dataframe with mSGP data for the year
}
\description{
Fetch mSGP
}
