% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{fetch_parcc}
\alias{fetch_parcc}
\title{Gets and cleans up a PARCC data file}
\usage{
fetch_parcc(end_year, grade_or_subj, subj, tidy = FALSE)
}
\arguments{
\item{end_year}{A school year. end_year is the end of the academic year - eg 2014-15
school year is end_year 2015. Valid values are 2015-2024.}

\item{grade_or_subj}{Grade level (eg 8) OR math subject code (eg ALG1, GEO, ALG2)}

\item{subj}{PARCC subject. c('ela' or 'math')}

\item{tidy}{Clean up the data frame to make it more compatible with
NJASK naming conventions and do some additional calculations? Default is FALSE.}
}
\value{
Processed PARCC/NJSLA dataframe
}
\description{
`fetch_parcc` is a wrapper around `get_raw_parcc` and `process_parcc`
that gets a parcc file and performs any cleanup.
}
\examples{
\dontrun{
# Get 2023 grade 4 math results
parcc_2023 <- fetch_parcc(2023, 4, "math")

# Get 2023 Algebra 1 results
alg1_2023 <- fetch_parcc(2023, "ALG1", "math")
}
}
