% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc.R
\name{fetch_parcc}
\alias{fetch_parcc}
\title{gets and cleans up a PARCC data file file}
\usage{
fetch_parcc(end_year, grade, subj)
}
\arguments{
\item{end_year}{a school year.  end_year is the end of the academic year - eg 2014-15
school year is end_year '2015'.  valid values are 2015}

\item{grade}{grade level}

\item{subj}{PARCC subject. c('ela' or 'math')}
}
\description{
\code{fetch_parcc} is a wrapper around \code{get_raw_parcc} and
\code{process_parcc} that gets a parcc file and performs any celanup.
}

