% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_pop.R
\name{fetch_reportcard_special_pop}
\alias{fetch_reportcard_special_pop}
\title{Fetch Special Population data}
\usage{
fetch_reportcard_special_pop(end_year)
}
\arguments{
\item{end_year}{ending academic year.  valid values are 2017, 2018, 2019}
}
\value{
data.frame with special population enrollment data
}
\description{
Fetch Special Population data
}
