% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped.R
\name{fetch_sped}
\alias{fetch_sped}
\title{Gets and cleans older SPED data}
\usage{
fetch_sped(end_year)
}
\arguments{
\item{end_year}{@inheritParams fetch_enr}
}
\value{
cleaned sped dataframe
}
\description{
Gets and cleans older SPED data
}
