% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spr.R
\name{fetch_violence_vandalism_hib}
\alias{fetch_violence_vandalism_hib}
\title{Fetch Violence/Vandalism/HIB Data}
\usage{
fetch_violence_vandalism_hib(end_year, level = "school")
}
\arguments{
\item{end_year}{A school year (2017-2024)}

\item{level}{One of "school" or "district"}
}
\value{
Data frame with incident counts
}
\description{
Downloads incident data from SPR database.
}
\examples{
\dontrun{
incidents <- fetch_violence_vandalism_hib(2024)
}
}
