% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation_aggs.R
\name{gcount_column_order}
\alias{gcount_column_order}
\title{Grad Count column order}
\usage{
gcount_column_order(df)
}
\arguments{
\item{df}{Processed grad count df}
}
\value{
Data frame with columns in correct order
}
\description{
Puts graduation count data frame columns in standard order.
}
\keyword{internal}
