% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgp.R
\name{get_and_process_msgp}
\alias{get_and_process_msgp}
\title{Get and Process mSGP data}
\usage{
get_and_process_msgp(end_year)
}
\arguments{
\item{end_year}{ending school year.  valid values are currently 2012-2018.}
}
\value{
df of msgp data, schoolwide (and district-wide and by grade level, if reported)
}
\description{
Get and Process mSGP data
}
