% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ell.R
\name{get_available_ell_years}
\alias{get_available_ell_years}
\title{Get the years for which NJ English Learner population data is available}
\usage{
get_available_ell_years()
}
\value{
integer vector of valid `end_year` values
}
\description{
Get the years for which NJ English Learner population data is available
}
\examples{
get_available_ell_years()
}
