% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_district_directory}
\alias{get_district_directory}
\title{Get NJ District Directory Data}
\usage{
get_district_directory()
}
\value{
dataframe of districts and associated metadata
}
\description{
Get NJ District Directory Data
}
