% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R, R/fetch_directory.R
\name{get_district_directory_url}
\alias{get_district_directory_url}
\title{Get district directory URL}
\usage{
get_district_directory_url()

get_district_directory_url()
}
\value{
Character string URL

Character string URL for district directory CSV
}
\description{
Get district directory URL

Get district directory download URL
}
\keyword{internal}
