% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R
\name{get_district_directory_url}
\alias{get_district_directory_url}
\title{Get district directory URL}
\usage{
get_district_directory_url()
}
\value{
Character string URL
}
\description{
Get district directory URL
}
\keyword{internal}
