% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{get_grad_rate}
\alias{get_grad_rate}
\title{Get NJ graduation rate data}
\usage{
get_grad_rate(end_year, methodology)
}
\arguments{
\item{end_year}{end of the academic year - 2011-2012 is 2012
valid values are 2011-present.}

\item{methodology}{character string specifying calculation methodology. One of "4 year" or "5 year"}
}
\value{
dataframe with the number of graduates per school and district
}
\description{
Get NJ graduation rate data
}
