% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enr.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{read a zipped excel fall enrollment file from the NJ state website}
\usage{
get_raw_enr(start_year)
}
\arguments{
\item{start_year}{a school year.  year is the start of the academic year - eg 2006-07
school year is year '2006'.  valid values are 1998-2015.}
}
\description{
\code{get_raw_enr} returns a data frame with a year's worth of fall school and
grade level enrollment data.
}

