% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Read a zipped Excel fall enrollment file from the NJ state website}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{A school year. Year is the end of the academic year - eg 2006-07
school year is year '2007'. Valid values are 1999-2025.}
}
\value{
Data frame with raw enrollment data
}
\description{
Read a zipped Excel fall enrollment file from the NJ state website
}
\keyword{internal}
