% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{get_raw_grad_file}
\alias{get_raw_grad_file}
\title{Get a raw graduation file from the NJ website}
\usage{
get_raw_grad_file(end_year, methodology = "4 year")
}
\arguments{
\item{end_year}{end of the academic year - eg 2006-07 is 2007.
valid values are 1998-present.}

\item{methodology}{one of c('4 year', '5 year')}
}
\value{
data.frame with raw data from state file
}
\description{
Get a raw graduation file from the NJ website
}
