% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_raw_njgpa}
\alias{get_raw_njgpa}
\title{Reads the raw NJGPA Excel files from the state website}
\usage{
get_raw_njgpa(end_year, subj)
}
\arguments{
\item{end_year}{A school year. Valid values are 2022-2024.}

\item{subj}{NJGPA subject. c('ela' or 'math')}
}
\value{
NJGPA dataframe
}
\description{
NJGPA (New Jersey Graduation Proficiency Assessment) is the graduation
requirement assessment introduced in 2022.
}
\keyword{internal}
