% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_raw_sla}
\alias{get_raw_sla}
\title{Reads the raw NJSLA Excel files from the state website}
\usage{
get_raw_sla(end_year, grade_or_subj, subj)
}
\arguments{
\item{end_year}{A school year. end_year is the end of the academic year - eg 2014-15
school year is end_year 2015. Valid values are 2015-2018.}

\item{grade_or_subj}{Grade level (eg 8) OR math subject code (eg ALG1, GEO, ALG2)}

\item{subj}{PARCC subject. c('ela' or 'math')}
}
\value{
NJSLA dataframe
}
\description{
Builds a URL and reads the xlsx file into a dataframe.
}
\keyword{internal}
