% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped.R
\name{get_raw_sped}
\alias{get_raw_sped}
\title{read Special ed excel files from the NJ state website}
\usage{
get_raw_sped(end_year)
}
\arguments{
\item{end_year}{A school year. Year is the end of the academic year - eg 2006-07
school year is year '2007'. Valid values are 1999-2025.}
}
\value{
a dataframe with special ed counts, etc.
}
\description{
read Special ed excel files from the NJ state website
}
