% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{get_raw_tges}
\alias{get_raw_tges}
\title{Get Raw Taxpayer's Guide to Educational Spending}
\usage{
get_raw_tges(end_year)
}
\arguments{
\item{end_year}{a school year.  end_year is the end of the academic year - eg 2016-17
school year is end_year 2017.  valid values are 1999-2017}
}
\value{
list of data frames
}
\description{
Get Raw Taxpayer's Guide to Educational Spending
}
\keyword{internal}
