% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{get_rc_databases}
\alias{get_rc_databases}
\title{Get multiple RC databases}
\usage{
get_rc_databases(end_year_vector = c(2003:2018))
}
\arguments{
\item{end_year_vector}{vector of years.  Current valid values are 2003 to 2018.}
}
\value{
a list of dataframes
}
\description{
Get multiple RC databases
}
