% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_school_directory}
\alias{get_school_directory}
\title{Get NJ School Directory Data}
\usage{
get_school_directory()
}
\value{
dataframe of schools and associated metadata
}
\description{
Get NJ School Directory Data
}
