% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R, R/fetch_directory.R
\name{get_school_directory_url}
\alias{get_school_directory_url}
\title{Get school directory URL}
\usage{
get_school_directory_url()

get_school_directory_url()
}
\value{
Character string URL

Character string URL for school directory CSV
}
\description{
Get school directory URL

Get school directory download URL
}
\keyword{internal}
