% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_config.R
\name{get_school_directory_url}
\alias{get_school_directory_url}
\title{Get school directory URL}
\usage{
get_school_directory_url()
}
\value{
Character string URL
}
\description{
Get school directory URL
}
\keyword{internal}
