% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card.R
\name{get_standalone_rc_database}
\alias{get_standalone_rc_database}
\title{Get Standalone Raw Report Card Database}
\usage{
get_standalone_rc_database(end_year)
}
\arguments{
\item{end_year}{a school year.  end_year is the end of the academic year - eg 2014-15
school year is end_year '2015'.  valid values are 2003 to 2016}
}
\value{
list of data frames
}
\description{
Get Standalone Raw Report Card Database
}
