% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{get_valid_sped_placement_years}
\alias{get_valid_sped_placement_years}
\title{Valid years for SPED placement / educational-environment data}
\usage{
get_valid_sped_placement_years()
}
\value{
integer vector of supported end years
}
\description{
Valid years for SPED placement / educational-environment data
}
\keyword{internal}
