% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_valid_years}
\alias{get_valid_years}
\title{Get valid years for a data type}
\usage{
get_valid_years(data_type)
}
\arguments{
\item{data_type}{Character string identifying the data type}
}
\value{
Integer vector of valid years
}
\description{
Get valid years for a data type
}
\examples{
get_valid_years("enrollment")
get_valid_years("parcc")
}
