% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{grad_file_group_cleanup}
\alias{grad_file_group_cleanup}
\title{grad file group cleanup}
\usage{
grad_file_group_cleanup(group)
}
\arguments{
\item{group}{column of group (subgroup) data from NJ grad file}
}
\value{
column cleaned up subgroup
}
\description{
grad file group cleanup
}
