% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation_aggs.R
\name{grate_column_order}
\alias{grate_column_order}
\title{Grad Rate column order}
\usage{
grate_column_order(df)
}
\arguments{
\item{df}{Processed grad rate df}
}
\value{
Data frame with columns in correct order
}
\description{
Puts graduation rate data frame columns in standard order.
}
\keyword{internal}
