% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{id_charter_hosts}
\alias{id_charter_hosts}
\title{Identify charter host districts}
\usage{
id_charter_hosts(df)
}
\arguments{
\item{df}{dataframe of NJ school data containing `district_id` column}
}
\value{
df with host district id and name for every matching record
}
\description{
Identify charter host districts
}
