% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation_aggs.R
\name{id_grad_aggs}
\alias{id_grad_aggs}
\title{Identify graduation aggregation levels}
\usage{
id_grad_aggs(df)
}
\arguments{
\item{df}{Graduation dataframe, output of tidy_grad_rate or tidy_grad_count}
}
\value{
data.frame with boolean aggregation flags
}
\description{
Adds boolean flags to identify state, county, district, and school level records.
}
