% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\name{is_charter_district}
\alias{is_charter_district}
\title{Check if a district is a charter}
\usage{
is_charter_district(county_id)
}
\arguments{
\item{county_id}{County ID}
}
\value{
Logical
}
\description{
Check if a district is a charter
}
\keyword{internal}
