% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R
\name{is_valid_year}
\alias{is_valid_year}
\title{Check if year is valid for a data type}
\usage{
is_valid_year(end_year, data_type)
}
\arguments{
\item{end_year}{The school year (end year)}

\item{data_type}{One of "enrollment", "parcc", "grate_4yr", "grate_5yr",
"gcount", "legacy_assess"}
}
\value{
Logical indicating if year is valid
}
\description{
Check if year is valid for a data type
}
\keyword{internal}
