% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cleaning.R
\name{kill_padformulas}
\alias{kill_padformulas}
\title{Kill Excel Formula Padding For Numeric Strings}
\usage{
kill_padformulas(x)
}
\arguments{
\item{x}{a vector with strings entered as formulas - eg ="01"}
}
\value{
a vector with normalized strings
}
\description{
Removes Excel formula padding (="01") from strings, leaving just
the numeric value.
}
