% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_calcs.R
\name{matric_column_order}
\alias{matric_column_order}
\title{Matriculation column order}
\usage{
matric_column_order(df)
}
\arguments{
\item{df}{Processed matriculation df}
}
\value{
Data frame with columns in correct order
}
\description{
Puts matriculation data frame columns in standard order.
}
