% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pad_leading}
\alias{pad_leading}
\title{pad leading digits}
\usage{
pad_leading(vector, digits)
}
\arguments{
\item{vector}{character vector}

\item{digits}{ensure exactly this many digits by leading zero-padding}
}
\value{
character vector
}
\description{
pad leading digits
}
