% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{parcc_column_order}
\alias{parcc_column_order}
\title{PARCC column order}
\usage{
parcc_column_order(df)
}
\arguments{
\item{df}{Tidied PARCC dataframe. Called as final step in fetch_parcc when tidy=TRUE.}
}
\value{
PARCC df with columns in coherent order
}
\description{
Puts PARCC dataframe columns in coherent order.
}
\keyword{internal}
