% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc_calcs.R
\name{parcc_perf_level_counts}
\alias{parcc_perf_level_counts}
\title{PARCC counts by performance level}
\usage{
parcc_perf_level_counts(df)
}
\arguments{
\item{df}{dataframe, output of fetch_parcc}
}
\value{
df with counts of students by performance level
}
\description{
PARCC counts by performance level
}
