% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enr.R
\name{process_enr}
\alias{process_enr}
\title{process a nj enrollment file}
\usage{
process_enr(df)
}
\arguments{
\item{df}{a enr data frame (eg output of \code{get_raw_enr})}
}
\description{
\code{process_enr} does cleanup of dataframes returned by \code{get_raw_enr}
}

