% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr_program}
\alias{process_enr_program}
\title{Join program code to program name}
\usage{
process_enr_program(df)
}
\arguments{
\item{df}{Cleaned enrollment file}
}
\value{
Data frame with program_name added
}
\description{
Decode the program name using the prog_codes lookup table.
}
\keyword{internal}
