% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\name{process_grad_count}
\alias{process_grad_count}
\title{Process Grad Count Data}
\usage{
process_grad_count(df, end_year)
}
\arguments{
\item{df}{Output of get_grad_count}

\item{end_year}{End of the academic year}
}
\value{
Data frame with composite subgroups
}
\description{
Creates composite subgroups like black (black_m + black_f).
}
\keyword{internal}
