% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\name{process_grad_rate}
\alias{process_grad_rate}
\title{Process Grad Rate}
\usage{
process_grad_rate(df, end_year, methodology)
}
\arguments{
\item{df}{Output of get_grad_rate}

\item{end_year}{Ending academic year}

\item{methodology}{One of c('4 year', '5 year')}
}
\value{
Data frame with normalized grad rate variables
}
\description{
Custom processing for grad rate data beyond generic process_grate.
}
\keyword{internal}
