% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{process_grad_rate}
\alias{process_grad_rate}
\title{Process Grad Rate}
\usage{
process_grad_rate(df, end_year, methodology)
}
\arguments{
\item{df}{output of get_grad_rate}

\item{end_year}{ending academic year}

\item{methodology}{one of c('4 year', '5 year')}

\item{description}{to the extent that fetch_grad_rate needs its own custom
logic above and beyond the generic process_grate, it will live here}
}
\value{
dataframe with normalized grad rate variables
}
\description{
Process Grad Rate
}
