% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grate.R
\name{process_grate}
\alias{process_grate}
\title{process grate}
\usage{
process_grate(df, end_year)
}
\arguments{
\item{df}{the output of get_raw_grate}

\item{end_year}{a school year.  year is the end of the academic year - eg 2006-07
school year is year '2007'.  valid values are 1998-2014.}
}
\description{
does cleanup of the grad rate ('grate') file
}

