% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_assess.R
\name{process_nj_assess}
\alias{process_nj_assess}
\title{process a nj assessment file}
\usage{
process_nj_assess(df, layout)
}
\arguments{
\item{df}{a raw NJASK, HSPA, or GEPA data frame (eg output of \code{get_raw_njask})
school year is year '2014'.  valid values are 2004-2014.}

\item{mask}{a vector indicating which columns are one implied decimal?
uses a layout file.}
}
\description{
\code{process_njask} does cleanup of the raw assessment file, primarily ensuring that
columns tagged as 'one implied' are displayed correctly#'
}

