% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assess.R
\name{process_nj_assess}
\alias{process_nj_assess}
\title{process a nj assessment file}
\usage{
process_nj_assess(df, layout)
}
\arguments{
\item{df}{a raw NJASK, HSPA, or GEPA data frame (eg output of \code{get_raw_njask})}

\item{layout}{which layout file to use to determine which columns are one implied 
decimal.}
}
\description{
\code{process_njask} does cleanup of the raw assessment file, primarily ensuring that 
columns tagged as 'one implied' are displayed correctly
}

