% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_parcc}
\alias{process_parcc}
\title{Process a raw PARCC data file}
\usage{
process_parcc(parcc_file, end_year, grade, subj)
}
\arguments{
\item{parcc_file}{Output of get_raw_parcc}

\item{end_year}{A school year. end_year is the end of the academic year - eg 2014-15
school year is end_year 2015. Valid values are 2015-2024.}

\item{grade}{Integer or character specifying grade level}

\item{subj}{PARCC subject. c('ela' or 'math')}
}
\value{
A tbl_df / data frame
}
\description{
All the logic needed to clean up the raw PARCC files.
}
\keyword{internal}
