% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_data.R
\docType{data}
\name{prog_codes}
\alias{prog_codes}
\title{program code data}
\format{
\describe{
\item{end_year}{end_year}
\item{program_code}{program_code}
\item{program_name}{program_name}
#' }
}
\source{
NJDOE website for each year of enrollment data (see link above)
}
\usage{
prog_codes
}
\description{
A compilation of all the "program codes" used by the state of NJ. example: 
http://www.nj.gov/education/data/enr/enr15/stat_doc.htm
}
\keyword{datasets}
