% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{sped_placement_cache_dir}
\alias{sped_placement_cache_dir}
\title{Directory holding cached SPED placement workbooks}
\usage{
sped_placement_cache_dir()
}
\value{
absolute path to the cache directory (created lazily on use)
}
\description{
Directory holding cached SPED placement workbooks
}
\keyword{internal}
