% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{split_enr_cols}
\alias{split_enr_cols}
\title{split enr columns}
\usage{
split_enr_cols(df)
}
\arguments{
\item{df}{a enr data frame (eg output of \code{get_raw_enr})}
}
\description{
splits enrollment columns that combine IDs and names (pre '09-10)
}

