% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{split_state_ed_environ_tables}
\alias{split_state_ed_environ_tables}
\title{Identify the next-table-divider row inside a stacked State sheet (2025)}
\usage{
split_state_ed_environ_tables(df)
}
\description{
Identify the next-table-divider row inside a stacked State sheet (2025)
}
\keyword{internal}
