% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_nj_assess.R
\name{standard_assess}
\alias{standard_assess}
\title{call the correct \code{fetch} function for normal assessment years}
\usage{
standard_assess(year, grade)
}
\arguments{
\item{year}{a school year.  year is the end of the academic year - eg 2013-14
school year is year '2014'.  valid values are 2004-2014.}

\item{grade}{a grade level.  valid values are 3,4,5,6,7,8}
}
\description{
for 2008-2014, this function will grab the NJASK for gr 3-8, and HSPA
for grade 11
}

