% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_constants.R
\docType{data}
\name{STATE_DISTRICT_ID}
\alias{STATE_DISTRICT_ID}
\title{State aggregate indicator codes}
\format{
An object of class \code{character} of length 1.
}
\usage{
STATE_DISTRICT_ID
}
\description{
State aggregate indicator codes
}
\keyword{internal}
