% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{tidy_enr}
\alias{tidy_enr}
\title{tidy enrollment data}
\usage{
tidy_enr(df)
}
\arguments{
\item{df}{a wide data.frame of processed enrollment data - eg output of \code{fetch_enr}}
}
\value{
a long data.frame of tidied enrollment data
}
\description{
tidy enrollment data
}
