% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grate.R
\name{tidy_grad_count}
\alias{tidy_grad_count}
\title{Tidy Grad Count}
\usage{
tidy_grad_count(df, end_year)
}
\arguments{
\item{df}{output of process_grad_count}

\item{end_year}{end of the academic year - eg 2006-07 is 2007.
valid values are 1998-present.}
}
\value{
data.frame with number of graduates
}
\description{
Tidy Grad Count
}
