% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch_nj_assess.R
\name{tidy_nj_assess}
\alias{tidy_nj_assess}
\title{tidies NJ assessment data}
\usage{
tidy_nj_assess(assess_name, df)
}
\arguments{
\item{assess_name}{NJASK, GEPA, HSPA}

\item{df}{a processed data frame (eg, output of process_njask)}
}
\description{
\code{tidy_nj_assess} is a utility/internal function that takes the somewhat messy/inconsistent
assessment headers and returns a tidy data frame.
}

