% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc.R
\name{tidy_parcc_subgroup}
\alias{tidy_parcc_subgroup}
\title{tidy parcc subgroup}
\usage{
tidy_parcc_subgroup(subgroup_vector)
}
\arguments{
\item{subgroup_vector}{subgroup column from parcc data file}
}
\value{
character vector with consistent subgroup names
}
\description{
tidy parcc subgroup
}
