% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{tidy_sped_placement_district_3_5}
\alias{tidy_sped_placement_district_3_5}
\title{Tidy the 3-5 District Counts sheet -- 2025}
\usage{
tidy_sped_placement_district_3_5(df)
}
\description{
Tidy the 3-5 District Counts sheet -- 2025
}
\keyword{internal}
