% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{tidy_sped_placement_district_5_21}
\alias{tidy_sped_placement_district_5_21}
\title{Tidy the District by Ed Environ sheet (school-age, 5-21) -- 2025}
\usage{
tidy_sped_placement_district_5_21(df)
}
\description{
Tidy the District by Ed Environ sheet (school-age, 5-21) -- 2025
}
\keyword{internal}
