% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sped_placement.R
\name{tidy_sped_placement_state}
\alias{tidy_sped_placement_state}
\title{Tidy the State by Ed Environ sheet (5-21 or 3-5)}
\usage{
tidy_sped_placement_state(df, age_group)
}
\arguments{
\item{df}{raw tibble from \code{get_raw_sped_placement(level = "state")}}

\item{age_group}{"5-21" or "3-5"}
}
\value{
tidy tibble in the same long shape as the district output
}
\description{
Tidy the State by Ed Environ sheet (5-21 or 3-5)
}
\keyword{internal}
