% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{unzipper}
\alias{unzipper}
\title{Download and unzip a file from URL}
\usage{
unzipper(url, file_pattern = "njschooldata")
}
\arguments{
\item{url}{character, target url}

\item{file_pattern}{character, stub to use for temporary file}
}
\value{
list of unzipped files in a temp directory
}
\description{
Downloads a zip file from a URL to a temporary location and extracts it.
}
\keyword{internal}
