% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ward_enr_aggs}
\alias{ward_enr_aggs}
\title{Aggregates enrollment data by ward}
\usage{
ward_enr_aggs(df)
}
\arguments{
\item{list_of_dfs}{output of \code{fetch_enr}}
}
\value{
A data frame of ward aggregations
}
\description{
Aggregates enrollment data by ward
}
