% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ward_parcc_aggs}
\alias{ward_parcc_aggs}
\title{Aggregates assessment data by ward}
\usage{
ward_parcc_aggs(list_of_dfs)
}
\arguments{
\item{list_of_dfs}{output of \code{fetch_all_parcc}, \code{fetch_parcc},
\code{calculate_agg_parcc_prof}}
}
\value{
A data frame of ward aggregations
}
\description{
Aggregates assessment data by ward
}
