% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_years.R, R/validation.R
\docType{data}
\name{year_ranges}
\alias{year_ranges}
\title{Valid year ranges for each data type}
\format{
An object of class \code{list} of length 12.
}
\usage{
year_ranges
}
\description{
Valid year ranges for each data type

Valid year ranges by data type
}
\keyword{internal}
